refactor(rest): split db.py into the ada.comms.rest.db package
The flat 4,340-line Postgres module becomes a package with one module
per table domain (pool, migrations, projects, audit_log, audit_queries,
audit_runs, audit_schedules, issue_bot, metrics, profiles, corpora,
catalogs, procedural_models, source_nodes, plugin_jobs, settings,
workers) plus a small shared helper module. The package __init__
re-exports every public and underscore-prefixed name, so
`from ada.comms.rest import db` / `import ada.comms.rest.db as db_module`
attribute access and test monkeypatching keep working unchanged, and
ada.topo_model.catalog's structural CatalogDbAdapter still matches.

No behaviour change; no new module-import-time dependencies.
