--- src/ada/comms/rest/db.py
@@ -1,4340 +1,0 @@
-"""Postgres layer for the multi-tenant REST viewer.
-
-Optional. When ``DATABASE_URL`` is empty the pool stays ``None`` and
-the API serves in shared-only mode: every authenticated user lands in
-the same shared bucket, no projects, no admin panel, no audit log.
-This keeps the helm chart genuinely "Postgres-optional" — small
-deployments don't need to stand up a database to use the viewer.
-
-Migrations are bundled SQL files under ``migrations/`` and applied at
-boot inside an advisory lock so a multi-replica rollout doesn't race.
-A row in ``schema_version`` records each applied file by stem name.
-
-Repository helpers expose the small set of queries the rest of the
-package needs (project list, membership check, audit log insert).
-asyncpg's pool is held on FastAPI's ``app.state.db_pool`` and pulled
-via the :func:`get_pool` accessor.
-"""
-
-from __future__ import annotations
-
-import datetime
-import hashlib
-import importlib.resources
-import json
-import re
-from dataclasses import dataclass
-from typing import Optional
-
-import asyncpg
-
-from ada.config import logger
-
-# Postgres advisory-lock id for the migration runner. asyncpg / PG
-# treat advisory keys as int8 (signed 64-bit), so the value must fit
-# in [-2^63, 2^63). Top nibble cleared so this stays a positive int.
-_MIGRATION_LOCK_ID = 0x0ADA0001_ADA00001
-
-
-@dataclass(frozen=True)
-class Project:
-    id: str
-    slug: str
-    name: str
-    role: str  # caller's role within this project
-
-
-async def init_pool(database_url: str) -> Optional[asyncpg.Pool]:
-    """Build a connection pool and apply pending migrations.
-
-    Returns ``None`` when ``database_url`` is empty so callers can
-    branch into shared-only mode without a try/except.
-
-    Retries connection failures with exponential backoff. The pod
-    can come up before kube-dns or before Postgres has finished
-    accepting connections, and a one-shot init that gives up on the
-    first ``socket.gaierror`` was leaving the API permanently in
-    shared-only mode (admin endpoints all returning 503) until
-    someone manually rolled the pod. Total budget ~60s — long enough
-    to ride out cold-start ordering, short enough that a real
-    misconfig still surfaces in the readiness probe.
-    """
-    if not database_url:
-        logger.info("db: DATABASE_URL not set — running in shared-only mode")
-        return None
-    import asyncio as _asyncio
-
-    delay = 1.0
-    deadline = 60.0
-    waited = 0.0
-    while True:
-        try:
-            pool = await asyncpg.create_pool(
-                dsn=database_url,
-                min_size=1,
-                max_size=10,
-                # Kill idle connections after 10 min — Postgres servers behind
-                # load balancers (PgBouncer, Garage's PG, managed Postgres) are
-                # finicky about long-idle conns.
-                max_inactive_connection_lifetime=600.0,
-            )
-            break
-        except (OSError, asyncpg.exceptions.PostgresError) as exc:
-            if waited >= deadline:
-                logger.error(
-                    "db: pool init still failing after %.0fs — giving up: %s",
-                    waited,
-                    exc,
-                )
-                raise
-            logger.warning(
-                "db: pool init failed (%s); retry in %.1fs (waited %.1fs/%.0fs)",
-                exc,
-                delay,
-                waited,
-                deadline,
-            )
-            await _asyncio.sleep(delay)
-            waited += delay
-            delay = min(delay * 1.6, 8.0)
-    try:
-        await _apply_migrations(pool)
-    except Exception:
-        await pool.close()
-        raise
-    logger.info("db: pool ready, migrations up-to-date")
-    return pool
-
-
-async def close_pool(pool: Optional[asyncpg.Pool]) -> None:
-    if pool is not None:
-        await pool.close()
-
-
-async def _apply_migrations(pool: asyncpg.Pool) -> None:
-    """Run any unapplied migrations under a Postgres advisory lock.
-
-    The lock is held only during apply; competing replicas wait, then
-    discover the migrations are already applied and become a no-op.
-    """
-    files = sorted(
-        p for p in importlib.resources.files("ada.comms.rest.migrations").iterdir() if p.name.endswith(".sql")
-    )
-
-    async with pool.acquire() as conn:
-        await conn.execute("SELECT pg_advisory_lock($1)", _MIGRATION_LOCK_ID)
-        try:
-            await conn.execute(
-                """
-                CREATE TABLE IF NOT EXISTS schema_version (
-                    version    TEXT PRIMARY KEY,
-                    applied_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
-                )
-                """
-            )
-            applied = {r["version"] for r in await conn.fetch("SELECT version FROM schema_version")}
-            for path in files:
-                version = path.stem
-                if version in applied:
-                    continue
-                logger.info("db: applying migration %s", version)
-                sql = path.read_text(encoding="utf-8")
-                async with conn.transaction():
-                    await conn.execute(sql)
-                    await conn.execute("INSERT INTO schema_version(version) VALUES ($1)", version)
-        finally:
-            await conn.execute("SELECT pg_advisory_unlock($1)", _MIGRATION_LOCK_ID)
-
-
-# ── Repository helpers ────────────────────────────────────────────────
-
-
-async def upsert_user(pool: asyncpg.Pool, sub: str, email: str, display_name: str) -> None:
-    """Lazy user upsert on first authenticated request. Bumps last_seen_at."""
-    await pool.execute(
-        """
-        INSERT INTO users (sub, email, display_name)
-        VALUES ($1, $2, $3)
-        ON CONFLICT (sub) DO UPDATE SET
-            email = EXCLUDED.email,
-            display_name = EXCLUDED.display_name,
-            last_seen_at = NOW()
-        """,
-        sub,
-        email or None,
-        display_name or None,
-    )
-
-
-async def list_user_projects(pool: asyncpg.Pool, user_sub: str) -> list[Project]:
-    rows = await pool.fetch(
-        """
-        SELECT p.id, p.slug, p.name, m.role
-        FROM project_members m
-        JOIN projects p ON p.id = m.project_id
-        WHERE m.user_sub = $1 AND p.archived_at IS NULL
-        ORDER BY p.name
-        """,
-        user_sub,
-    )
-    return [Project(id=str(r["id"]), slug=r["slug"], name=r["name"], role=r["role"]) for r in rows]
-
-
-async def is_project_member(pool: asyncpg.Pool, project_id: str, user_sub: str) -> bool:
-    row = await pool.fetchrow(
-        """
-        SELECT 1
-        FROM project_members m
-        JOIN projects p ON p.id = m.project_id
-        WHERE m.user_sub = $1 AND p.id = $2 AND p.archived_at IS NULL
-        """,
-        user_sub,
-        project_id,
-    )
-    return row is not None
-
-
-async def insert_audit(
-    pool: asyncpg.Pool,
-    *,
-    user_sub: str | None,
-    scope_kind: str,
-    scope_id: str | None,
-    action: str,
-    key: str | None = None,
-    target_format: str | None = None,
-    status: str | None = None,
-    error: str | None = None,
-    duration_ms: int | None = None,
-    job_id: str | None = None,
-    traceback: str | None = None,
-    audit_run_id: str | None = None,
-    worker_image_tag: str | None = None,
-    read_bytes: int | None = None,
-    write_bytes: int | None = None,
-    peak_rss_kb: int | None = None,
-    client_metrics: dict | None = None,
-    failure_key: str | None = None,
-) -> None:
-    """Insert one audit_log row.
-
-    ``audit_run_id`` links the row back to an admin-triggered
-    regression sweep (M1 audit panel). NULL for every user-driven
-    convert / upload / download row — only the audit dispatcher
-    populates it.
-
-    When the row both carries an ``audit_run_id`` AND lands in a
-    terminal status (``done`` / ``ok`` / ``error`` / ``cancelled``
-    / ``skipped``), the matching counter on ``audit_runs`` is bumped
-    in the same transaction. This is the cached-cell path of the
-    dispatcher — derived blobs that already exist get audited as
-    ``done`` without enqueueing a job, and the run's ok counter
-    needs to advance immediately so the math closes against the
-    total.
-    """
-
-    counter_col = _AUDIT_RUN_COUNTER_FOR_STATUS.get(status) if audit_run_id is not None and status is not None else None
-    if counter_col is None:
-        # Hot path — single INSERT, no transaction overhead. Covers
-        # every user-driven action and the audit dispatcher's
-        # ``status='queued'`` enqueue audit.
-        await pool.execute(
-            """
-            INSERT INTO audit_log
-                (user_sub, scope_kind, scope_id, action, key,
-                 target_format, status, error, duration_ms, job_id,
-                 traceback, audit_run_id, worker_image_tag,
-                 read_bytes, write_bytes, peak_rss_kb, client_metrics,
-                 failure_key)
-            VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13,
-                    $14, $15, $16, $17::jsonb, $18)
-            """,
-            user_sub,
-            scope_kind,
-            scope_id,
-            action,
-            key,
-            target_format,
-            status,
-            error,
-            duration_ms,
-            job_id,
-            traceback,
-            audit_run_id,
-            worker_image_tag,
-            read_bytes,
-            write_bytes,
-            peak_rss_kb,
-            json.dumps(client_metrics) if client_metrics is not None else None,
-            failure_key,
-        )
-        return
-
-    # Audit-dispatcher cached-cell path: insert + counter bump in one
-    # transaction so a crash between the two can't leave a row counted
-    # by the audit grid but not by the run total.
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            await conn.execute(
-                """
-                INSERT INTO audit_log
-                    (user_sub, scope_kind, scope_id, action, key,
-                     target_format, status, error, duration_ms, job_id,
-                     traceback, audit_run_id, worker_image_tag,
-                     read_bytes, write_bytes, peak_rss_kb, client_metrics,
-                     failure_key)
-                VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13,
-                        $14, $15, $16, $17::jsonb, $18)
-                """,
-                user_sub,
-                scope_kind,
-                scope_id,
-                action,
-                key,
-                target_format,
-                status,
-                error,
-                duration_ms,
-                job_id,
-                traceback,
-                audit_run_id,
-                worker_image_tag,
-                read_bytes,
-                write_bytes,
-                peak_rss_kb,
-                json.dumps(client_metrics) if client_metrics is not None else None,
-                failure_key,
-            )
-            await _bump_audit_run_counter(conn, audit_run_id, status)
-
-
-async def cancel_audit_by_job(
-    pool: asyncpg.Pool,
-    *,
-    job_id: str,
-    user_sub: str,
-) -> bool:
-    """Mark a queued or running job as cancelled. Caller must own it.
-
-    Returns ``True`` if a row was updated, ``False`` otherwise (row
-    missing, status already terminal, or owned by someone else).
-    Filtering on ``status`` lets the SQL itself enforce the "only
-    cancel queued/running" rule so a race between two cancel clicks
-    can't mark a done job cancelled retroactively.
-    """
-    result = await pool.execute(
-        """
-        UPDATE audit_log
-        SET status = 'cancelled',
-            error = COALESCE(error, 'cancelled by user')
-        WHERE job_id = $1
-          AND user_sub = $2
-          AND status IN ('queued', 'running')
-        """,
-        job_id,
-        user_sub,
-    )
-    return result.endswith(" 1")
-
-
-async def admin_cancel_audit_by_job(
-    pool: asyncpg.Pool,
-    *,
-    job_id: str,
-    reason: str = "cancelled by an administrator",
-) -> bool:
-    """Cancel a job REGARDLESS of who started it. Admin-gated at the route.
-
-    The same UPDATE as :func:`cancel_audit_by_job` without the ownership
-    filter, and deliberately a separate function rather than an optional
-    ``user_sub=None``: a default that silently disables the owner check is one
-    forgotten keyword away from letting any user cancel anybody's work. Two
-    names, two call sites, no way to reach this one by omission.
-
-    The ``status IN ('queued','running')`` filter is kept for the same reason
-    it exists there -- a terminal row must not be rewritten retroactively, and
-    that rule does not relax for an admin.
-    """
-    result = await pool.execute(
-        """
-        UPDATE audit_log
-        SET status = 'cancelled',
-            error = COALESCE(error, $2)
-        WHERE job_id = $1
-          AND status IN ('queued', 'running')
-        """,
-        job_id,
-        reason,
-    )
-    return result.endswith(" 1")
-
-
-async def audit_is_cancelled(pool: asyncpg.Pool, job_id: str) -> bool:
-    """True once the job's audit row has been flipped to ``cancelled`` (the cancel
-    endpoint's source of truth — the KV status is overwritten by worker progress
-    writes, so it can't be trusted). Polled by the conversion watchdog so an
-    actively-running conversion is actually reaped instead of running to completion."""
-    row = await pool.fetchrow(
-        "SELECT 1 FROM audit_log WHERE job_id = $1 AND status = 'cancelled' LIMIT 1",
-        job_id,
-    )
-    return row is not None
-
-
-async def get_audit_owner_by_job(pool: asyncpg.Pool, job_id: str) -> dict | None:
-    """Return ``{user_sub, scope_kind, scope_id, audit_run_id, status}``
-    for the audit row tied to ``job_id``, or ``None`` if absent.
-
-    Used by the browser-driven (WASM) audit-update endpoint to verify
-    the caller owns the row before patching it — the worker's
-    ``update_audit_by_job`` has no such guard because only the worker
-    that owns the NATS job can reach it, whereas the local endpoint is
-    callable by any authenticated browser.
-    """
-    row = await pool.fetchrow(
-        """
-        SELECT user_sub, scope_kind, scope_id, audit_run_id, status
-        FROM audit_log
-        WHERE job_id = $1
-        ORDER BY id DESC
-        LIMIT 1
-        """,
-        job_id,
-    )
-    return dict(row) if row is not None else None
-
-
-async def mark_audit_running(
-    pool: asyncpg.Pool,
-    *,
-    job_id: str,
-    worker_image_tag: str | None = None,
-) -> None:
-    """Flip the audit_log row for a job into ``status='running'`` and stamp
-    ``started_at`` so the admin's "current cell" query has a fresh row.
-
-    Without this hop the row goes straight queued → done and the audit toast's
-    ORDER BY can't tell which queued cell is actually being worked on — so the
-    same one stays on screen for the whole sweep.
-
-    THIS USED TO OVERWRITE ``ts``, and that is why queue wait could not be
-    measured after the fact: ``ts`` was the enqueue time while a row was queued
-    and the START time from the moment it was picked up, so the interval between
-    the two — the wait — was destroyed by the very hop that ended it. It also
-    made ``ts`` mean two different things depending on a column next to it,
-    which is a trap for anything reading the table.
-
-    ``ts`` is now the enqueue time for the row's whole life, and ``started_at``
-    is the pickup. Wait is the difference. Best-effort: a DB hiccup must never
-    break job processing.
-    """
-    if pool is None:
-        return
-    try:
-        await pool.execute(
-            """
-            UPDATE audit_log
-            SET status = 'running',
-                started_at = NOW(),
-                worker_image_tag = COALESCE($2, worker_image_tag)
-            WHERE job_id = $1
-              AND status = 'queued'
-            """,
-            job_id,
-            worker_image_tag,
-        )
-    except Exception:
-        logger.exception("worker: mark_audit_running failed for job %s", job_id)
-
-
-async def update_audit_by_job(
-    pool: asyncpg.Pool,
-    *,
-    job_id: str,
-    status: str,
-    error: str | None = None,
-    duration_ms: int | None = None,
-    traceback: str | None = None,
-    cpu_user_ms: int | None = None,
-    cpu_sys_ms: int | None = None,
-    peak_rss_kb: int | None = None,
-    read_bytes: int | None = None,
-    write_bytes: int | None = None,
-    profile_key: str | None = None,
-    log_key: str | None = None,
-    worker_image_tag: str | None = None,
-    convert_meta: dict | None = None,
-    failure_key: str | None = None,
-) -> None:
-    """Patch the audit row tied to a queue job with its final outcome.
-
-    No-op when the row is missing (job predates the migration, or the
-    enqueue-time audit insert failed). COALESCE preserves any existing
-    column when the caller passes None.
-
-    When the row carries an ``audit_run_id`` (M1 audit panel), the
-    matching counter on ``audit_runs`` (``ok`` / ``failed`` /
-    ``skipped``) is bumped in the same transaction. The run flips to
-    ``status='finished'`` when ``ok + failed + skipped`` reaches
-    ``total``. Regular user-driven jobs (audit_run_id IS NULL) take
-    the original single-table path with no overhead.
-    """
-    # Single transaction so the per-row write + the run-counter bump
-    # never split — a worker restart between the two would otherwise
-    # leak a job from the counters and the run would never finish.
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            updated = await conn.fetchrow(
-                """
-                UPDATE audit_log
-                SET status = $2,
-                    error = COALESCE($3, error),
-                    duration_ms = COALESCE($4, duration_ms),
-                    traceback = COALESCE($5, traceback),
-                    cpu_user_ms = COALESCE($6, cpu_user_ms),
-                    cpu_sys_ms = COALESCE($7, cpu_sys_ms),
-                    peak_rss_kb = COALESCE($8, peak_rss_kb),
-                    read_bytes = COALESCE($9, read_bytes),
-                    write_bytes = COALESCE($10, write_bytes),
-                    profile_key = COALESCE($11, profile_key),
-                    worker_image_tag = COALESCE($12, worker_image_tag),
-                    convert_meta = COALESCE($13, convert_meta),
-                    log_key = COALESCE($14, log_key),
-                    failure_key = COALESCE($15, failure_key)
-                WHERE job_id = $1
-                RETURNING audit_run_id
-                """,
-                job_id,
-                status,
-                error,
-                duration_ms,
-                traceback,
-                cpu_user_ms,
-                cpu_sys_ms,
-                peak_rss_kb,
-                read_bytes,
-                write_bytes,
-                profile_key,
-                worker_image_tag,
-                json.dumps(convert_meta) if convert_meta is not None else None,
-                log_key,
-                failure_key,
-            )
-            if updated is None or updated["audit_run_id"] is None:
-                return
-            await _bump_audit_run_counter(conn, updated["audit_run_id"], status)
-
-
-# Map terminal job-status values to the audit_runs counter column they
-# bump. Statuses outside this set (``running`` / ``queued``) don't
-# advance any counter — only terminal transitions do.
-_AUDIT_RUN_COUNTER_FOR_STATUS = {
-    "done": "ok",
-    "ok": "ok",
-    "error": "failed",
-    "failed": "failed",
-    "cancelled": "skipped",
-    "skipped": "skipped",
-}
-
-
-async def _bump_audit_run_counter(
-    conn: asyncpg.Connection,
-    run_id,
-    terminal_status: str,
-) -> None:
-    """Increment one of the run's terminal counters and finish the
-    run when all enqueued jobs have landed. Connection-bound (not
-    pool-bound) so the caller can run this inside the same
-    transaction as the audit_log UPDATE.
-
-    ``status='aborted'`` (an admin pressed Cancel) is preserved —
-    late worker completions arriving after the abort still bump
-    counters for diagnostics, but the run never auto-flips back to
-    ``finished``.
-    """
-    column = _AUDIT_RUN_COUNTER_FOR_STATUS.get(terminal_status)
-    if column is None:
-        # Transient state (``running``); nothing to bump yet.
-        return
-    # The SET clause uses dynamic column interpolation but ``column``
-    # is constrained to a closed allowlist above, so f-string here is
-    # safe — no caller-supplied SQL surface.
-    await conn.execute(
-        f"""
-        UPDATE audit_runs
-        SET {column} = {column} + 1,
-            finished_at = CASE
-                WHEN status = 'aborted' THEN finished_at
-                WHEN ok + failed + skipped + 1 >= total
-                  THEN COALESCE(finished_at, NOW())
-                ELSE finished_at
-            END,
-            status = CASE
-                WHEN status = 'aborted' THEN 'aborted'
-                WHEN ok + failed + skipped + 1 >= total
-                  THEN 'finished'
-                ELSE status
-            END
-        WHERE id = $1
-        """,
-        run_id,
-    )
-
-
-async def reset_audit_cell_for_rerun(
-    pool: asyncpg.Pool,
-    run_id: str,
-    key: str,
-    target_format: str,
-    new_job_id: str,
-) -> bool:
-    """Re-point a single audit cell at a fresh conversion job so it can be
-    re-run in place, keeping the run's counters and total correct.
-
-    One row exists per cell in a run (the dispatcher inserts it ``queued`` and
-    the worker updates it by ``job_id``). To re-run one cell we: undo the
-    counter its prior terminal status bumped (so ``failed`` drops when a
-    previously-failed cell is retried), clear its result fields, attach the new
-    ``job_id`` and set it back to ``queued``, and reopen the run. The worker's
-    normal completion path (``update_audit_by_job`` → the counter bump)
-    then re-increments and re-finishes the run exactly as for a first run.
-
-    Returns False if the cell isn't part of the run (nothing to re-run)."""
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            row = await conn.fetchrow(
-                """
-                SELECT id, status FROM audit_log
-                WHERE audit_run_id = $1 AND key = $2 AND target_format = $3
-                ORDER BY id DESC
-                LIMIT 1
-                """,
-                run_id,
-                key,
-                target_format,
-            )
-            if row is None:
-                return False
-            prior_col = _AUDIT_RUN_COUNTER_FOR_STATUS.get(row["status"] or "")
-            await conn.execute(
-                """
-                UPDATE audit_log
-                SET job_id = $2, status = 'queued', error = NULL, traceback = NULL,
-                    duration_ms = NULL, cpu_user_ms = NULL, cpu_sys_ms = NULL,
-                    peak_rss_kb = NULL, read_bytes = NULL, write_bytes = NULL,
-                    profile_key = NULL, log_key = NULL, ts = NOW()
-                WHERE id = $1
-                """,
-                row["id"],
-                new_job_id,
-            )
-            # Undo the prior terminal counter (if any) and reopen the run. The
-            # column name comes from a closed allowlist so the f-string is safe.
-            dec = f"{prior_col} = GREATEST({prior_col} - 1, 0)," if prior_col else ""
-            await conn.execute(
-                f"""
-                UPDATE audit_runs
-                SET {dec}
-                    -- Fold the idle gap since this run last finished into idle_ms
-                    -- (same as extend_audit_run_total). Then when the re-run cell
-                    -- lands and re-finishes the run, the wall-clock duration grows
-                    -- ONLY by the re-run's own delta, not the days-long gap since
-                    -- the original run. RHS reads the pre-update finished_at.
-                    idle_ms = idle_ms + CASE
-                        WHEN status <> 'aborted' AND finished_at IS NOT NULL
-                        THEN GREATEST(0, (EXTRACT(EPOCH FROM (NOW() - finished_at)) * 1000)::bigint)
-                        ELSE 0 END,
-                    status = CASE WHEN status = 'aborted' THEN 'aborted' ELSE 'running' END,
-                    finished_at = CASE WHEN status = 'aborted' THEN finished_at ELSE NULL END
-                WHERE id = $1
-                """,
-                run_id,
-            )
-            return True
-
-
-async def active_audit_summary(pool: asyncpg.Pool) -> dict:
-    """Aggregate state of currently-``running`` audit runs.
-
-    Used by the bottom-right viewer toast to show an ambient
-    "N audit runs · M cells pending" badge that links into the
-    admin panel. Plus the most recently-touched in-flight cell
-    (status ``running`` preferred, falls back to ``queued``) so
-    the operator sees what's actually converting right now —
-    handy on a force-rebuild measurement run where every cell
-    actually executes.
-
-    Returns:
-        {"running_runs": int, "pending_cells": int,
-         "current_cell": {key, target_format, status,
-                          started_at, elapsed_ms} | None}
-    """
-    counts_row = await pool.fetchrow(
-        """
-        SELECT
-            COUNT(*) AS running_runs,
-            COALESCE(SUM(GREATEST(total - ok - failed - skipped, 0)), 0)
-                AS pending_cells
-        FROM audit_runs
-        WHERE status = 'running'
-        """
-    )
-    # Only surface rows where the worker has actually picked the
-    # cell up — i.e. ``audit_log.status = 'running'`` (set by
-    # ``mark_audit_running`` at the start of ``_process_one``).
-    # Queued rows don't reflect what's converting right now; the
-    # dispatcher inserts them all with the same ts at the start of
-    # the sweep, so picking the "most recent queued" gives a row
-    # that doesn't change as the sweep advances. Caller sees no
-    # current_cell between cells (a few hundred ms) — that's the
-    # honest answer, and the toast hides the line there.
-    #
-    # Ordered by ``started_at``, which is what "most recently picked up" means
-    # now that ``ts`` stays at enqueue. COALESCE keeps a mid-rollout row written
-    # by an older worker (started_at NULL, ts stamped at pickup) in the right
-    # place instead of sorting it to the bottom for the length of the sweep.
-    current_row = await pool.fetchrow(
-        """
-        SELECT al.key, al.target_format, al.status,
-               COALESCE(al.started_at, al.ts) AS ts
-        FROM audit_log al
-        JOIN audit_runs ar ON ar.id = al.audit_run_id
-        WHERE ar.status = 'running'
-          AND al.status = 'running'
-        ORDER BY COALESCE(al.started_at, al.ts) DESC
-        LIMIT 1
-        """
-    )
-    current_cell = None
-    if current_row is not None:
-        ts = current_row["ts"]
-        elapsed_ms = None
-        if ts is not None:
-            from datetime import datetime, timezone
-
-            elapsed_ms = int((datetime.now(timezone.utc) - ts).total_seconds() * 1000)
-        current_cell = {
-            "key": current_row["key"],
-            "target_format": current_row["target_format"],
-            "status": current_row["status"],
-            "started_at": ts.isoformat() if ts else None,
-            "elapsed_ms": elapsed_ms,
-        }
-    return {
-        "running_runs": int(counts_row["running_runs"] or 0),
-        "pending_cells": int(counts_row["pending_cells"] or 0),
-        "current_cell": current_cell,
-    }
-
-
-async def abort_audit_run(
-    pool: asyncpg.Pool,
-    run_id: str,
-) -> dict | None:
-    """Stop a running audit. Sets the run's status to ``'aborted'``
-    and cancels every queued / running child audit_log row in the
-    same transaction so the per-cell grid shows where the run was
-    when it died (the rows that already finished keep their
-    terminal status — we don't rewrite history)."""
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            run = await conn.fetchrow(
-                """
-                UPDATE audit_runs
-                SET status = 'aborted',
-                    finished_at = NOW()
-                WHERE id = $1
-                  AND status = 'running'
-                RETURNING id, scope, worker_pool, trigger, started_at,
-                          finished_at, status, note, total, ok, failed,
-                          skipped, created_by, issue_bot_status,
-                          issue_bot_last_error, issue_bot_synced_at
-                """,
-                run_id,
-            )
-            if run is None:
-                return None
-            # Cancel still-queued / still-running children. Counts
-            # them as ``skipped`` so the run's ok+failed+skipped
-            # matches total once late completions stop landing.
-            cancelled_rows = await conn.fetch(
-                """
-                UPDATE audit_log
-                SET status = 'cancelled',
-                    error = COALESCE(error, 'audit run aborted')
-                WHERE audit_run_id = $1
-                  AND status IN ('queued', 'running')
-                RETURNING id, job_id
-                """,
-                run_id,
-            )
-            n_cancel = len(cancelled_rows)
-            if n_cancel > 0:
-                await conn.execute(
-                    """
-                    UPDATE audit_runs
-                    SET skipped = skipped + $2
-                    WHERE id = $1
-                    """,
-                    run_id,
-                    n_cancel,
-                )
-    result = _audit_run_row(run)
-    # job_ids of the cells we just cancelled — the caller purges their still-queued
-    # JetStream messages so the worker never pulls/processes a doomed cell.
-    result["cancelled_job_ids"] = [r["job_id"] for r in cancelled_rows if r["job_id"]]
-    return result
-
-
-async def append_metrics_sample_by_job(
-    pool: asyncpg.Pool,
-    *,
-    job_id: str,
-    sample: dict,
-) -> None:
-    """Append one heartbeat sample to ``audit_log.metrics_samples``.
-
-    The column is JSONB initialised to NULL when the row is created;
-    we coalesce to ``[]`` before appending so the first heartbeat
-    on a fresh row still produces a valid array. Best-effort —
-    callers swallow exceptions because losing one heartbeat shouldn't
-    fail the conversion.
-    """
-    await pool.execute(
-        """
-        UPDATE audit_log
-        SET metrics_samples = COALESCE(metrics_samples, '[]'::jsonb) || $2::jsonb
-        WHERE job_id = $1
-        """,
-        job_id,
-        json.dumps(sample),
-    )
-
-
-# ── Admin queries ────────────────────────────────────────────────────
-
-
-_RELATIVE_BOUND = re.compile(r"^(\d+)\s*([smhdw])$", re.IGNORECASE)
-_RELATIVE_UNIT = {
-    "s": "seconds",
-    "m": "minutes",
-    "h": "hours",
-    "d": "days",
-    "w": "weeks",
-}
-
-
-def parse_audit_time_bound(value: str | None, *, now: datetime.datetime | None = None) -> datetime.datetime | None:
-    """A time bound for the audit filter: ``"6h"``, or an ISO-8601 instant.
-
-    Relative forms are resolved HERE, against the server's clock, rather than
-    the browser computing an absolute instant and sending that. A workstation
-    whose clock is a few minutes fast would otherwise silently drop rows from a
-    "last 5 minutes" view and show a window that never existed — and the
-    narrower the range, the worse the error, which is exactly backwards.
-
-    Absolute ISO input stays absolute: for a custom range the operator picked
-    two real instants, and skew is irrelevant to what they meant.
-
-    Empty / None means unbounded. Anything unparseable raises ``ValueError`` so
-    the caller can answer 400 rather than quietly returning all of history.
-    """
-    raw = (value or "").strip()
-    if not raw:
-        return None
-
-    m = _RELATIVE_BOUND.match(raw)
-    if m:
-        amount, unit = int(m.group(1)), m.group(2).lower()
-        base = now or datetime.datetime.now(datetime.timezone.utc)
-        return base - datetime.timedelta(**{_RELATIVE_UNIT[unit]: amount})
-
-    iso = raw.replace("Z", "+00:00") if raw.endswith("Z") else raw
-    try:
-        parsed = datetime.datetime.fromisoformat(iso)
-    except ValueError as exc:
-        raise ValueError(f"not a duration (e.g. '6h') or an ISO-8601 instant: {value!r}") from exc
-    # A naive instant is taken as UTC: the column is timestamptz, and comparing
-    # it against a naive value raises rather than guessing a zone.
-    if parsed.tzinfo is None:
-        parsed = parsed.replace(tzinfo=datetime.timezone.utc)
-    return parsed
-
-
-def _audit_predicates(
-    *,
-    user_sub: str | None = None,
-    scope_kind: str | None = None,
-    scope_id: str | None = None,
-    action: str | None = None,
-    target_format: str | None = None,
-    statuses: list[str] | None = None,
-    key_like: str | None = None,
-    since: datetime.datetime | None = None,
-    until: datetime.datetime | None = None,
-    before_id: int | None = None,
-    exclude_audit_dispatched: bool = False,
-) -> tuple[list[str], list]:
-    """Build the shared ``audit_log`` WHERE fragments and their arguments.
-
-    Extracted so ``list_audit`` and ``summarize_audit`` cannot drift: the
-    admin Audit tab shows a summary and a log side by side under ONE filter,
-    and a predicate honoured by one but not the other reads as a counting bug
-    rather than as the mismatch it is. Placeholders are numbered from the
-    running length of ``args``, so a caller may append its own (a LIMIT, say)
-    afterwards.
-    """
-    where: list[str] = []
-    args: list = []
-    if user_sub:
-        args.append(user_sub)
-        where.append(f"user_sub = ${len(args)}")
-    if scope_kind:
-        args.append(scope_kind)
-        where.append(f"scope_kind = ${len(args)}")
-    if scope_id:
-        args.append(scope_id)
-        where.append(f"scope_id = ${len(args)}")
-    if action:
-        args.append(action)
-        where.append(f"action = ${len(args)}")
-    if target_format:
-        args.append(target_format)
-        where.append(f"target_format = ${len(args)}")
-    if statuses:
-        args.append(statuses)
-        where.append(f"status = ANY(${len(args)})")
-    if key_like:
-        args.append(f"%{key_like}%")
-        where.append(f"key ILIKE ${len(args)}")
-    # Bounded on ``ts``, which carries a DESC btree index, so narrowing the
-    # window makes both the log and the summary cheaper rather than dearer.
-    if since is not None:
-        args.append(since)
-        where.append(f"ts >= ${len(args)}")
-    if until is not None:
-        args.append(until)
-        where.append(f"ts <= ${len(args)}")
-    if before_id is not None:
-        args.append(before_id)
-        where.append(f"id < ${len(args)}")
-    if exclude_audit_dispatched:
-        where.append("audit_run_id IS NULL")
-    return where, args
-
-
-async def list_audit(
-    pool: asyncpg.Pool,
-    *,
-    user_sub: str | None = None,
-    scope_kind: str | None = None,
-    scope_id: str | None = None,
-    action: str | None = None,
-    target_format: str | None = None,
-    statuses: list[str] | None = None,
-    key_like: str | None = None,
-    since: datetime.datetime | None = None,
-    until: datetime.datetime | None = None,
-    limit: int = 100,
-    before_id: int | None = None,
-    exclude_audit_dispatched: bool = False,
-) -> list[dict]:
-    """Reverse-chronological audit_log scan, optionally filtered.
-
-    ``key_like`` is a case-insensitive substring filter on the source ``key``
-    (the filepath/filename), so the admin audit log can be narrowed to one file
-    or folder (``%term%`` ILIKE).
-
-    Pagination is keyset-style on ``id`` (the BIGSERIAL primary key) —
-    pass the smallest id from the previous page as ``before_id``. id
-    monotonicity matches ``ts`` ordering and avoids the offset-based
-    "page drift" surprise when new rows arrive between requests.
-
-    ``statuses`` filters by the job's terminal/transient state (e.g.
-    ``["queued", "running"]`` for the user-facing "my in-flight jobs"
-    view). Empty list / None disables the filter.
-
-    ``exclude_audit_dispatched`` filters out cells emitted by the
-    admin audit sweep (``audit_run_id IS NOT NULL``). The user-
-    facing /my-jobs view sets this so a 453-cell sweep doesn't
-    flood the bottom-right toast — the Audit Runs admin tab is
-    the proper surface for that work.
-    """
-    where, args = _audit_predicates(
-        user_sub=user_sub,
-        scope_kind=scope_kind,
-        scope_id=scope_id,
-        action=action,
-        target_format=target_format,
-        statuses=statuses,
-        key_like=key_like,
-        since=since,
-        until=until,
-        before_id=before_id,
-        exclude_audit_dispatched=exclude_audit_dispatched,
-    )
-    args.append(min(max(limit, 1), 500))
-    sql = (
-        "SELECT id, ts, started_at, user_sub, scope_kind, scope_id, action, key,"
-        " target_format, status, error, duration_ms, traceback,"
-        " cpu_user_ms, cpu_sys_ms, peak_rss_kb, read_bytes, write_bytes,"
-        " profile_key, log_key, job_id, audit_run_id, worker_image_tag, convert_meta,"
-        " issue_bot_status, issue_bot_synced_at, issue_bot_last_error,"
-        " client_metrics->>'device_id' AS device_id,"
-        " u.email AS user_email, u.display_name AS user_display_name"
-        " FROM audit_log LEFT JOIN users u ON u.sub = audit_log.user_sub"
-    )
-    if where:
-        sql += " WHERE " + " AND ".join(where)
-    sql += f" ORDER BY id DESC LIMIT ${len(args)}"
-    rows = await pool.fetch(sql, *args)
-    return [
-        {
-            "id": r["id"],
-            "ts": r["ts"].isoformat() if r["ts"] is not None else None,
-            "started_at": r["started_at"].isoformat() if r["started_at"] is not None else None,
-            "user_sub": r["user_sub"],
-            "user_email": r["user_email"],
-            "user_display_name": r["user_display_name"],
-            "scope_kind": r["scope_kind"],
-            "scope_id": r["scope_id"],
-            "action": r["action"],
-            "key": r["key"],
-            "target_format": r["target_format"],
-            "status": r["status"],
-            "error": r["error"],
-            "duration_ms": r["duration_ms"],
-            "traceback": r["traceback"],
-            "cpu_user_ms": r["cpu_user_ms"],
-            "cpu_sys_ms": r["cpu_sys_ms"],
-            "peak_rss_kb": r["peak_rss_kb"],
-            "read_bytes": r["read_bytes"],
-            "write_bytes": r["write_bytes"],
-            "profile_key": r["profile_key"],
-            "log_key": r["log_key"],
-            "job_id": r["job_id"],
-            "audit_run_id": str(r["audit_run_id"]) if r["audit_run_id"] else None,
-            "worker_image_tag": r["worker_image_tag"],
-            "convert_meta": _loads_jsonb(r["convert_meta"]),
-            "issue_bot_status": r["issue_bot_status"],
-            "issue_bot_synced_at": (r["issue_bot_synced_at"].isoformat() if r["issue_bot_synced_at"] else None),
-            "issue_bot_last_error": r["issue_bot_last_error"],
-            "device_id": r["device_id"],
-        }
-        for r in rows
-    ]
-
-
-async def summarize_audit(
-    pool: asyncpg.Pool,
-    *,
-    user_sub: str | None = None,
-    scope_kind: str | None = None,
-    scope_id: str | None = None,
-    action: str | None = None,
-    target_format: str | None = None,
-    key_like: str | None = None,
-    since: datetime.datetime | None = None,
-    until: datetime.datetime | None = None,
-    reason_limit: int = 10,
-) -> dict:
-    """Aggregate counts for the Audit tab's Overview, under the same filter
-    the log uses.
-
-    NOTE THE MISSING ``statuses`` PARAMETER — it is deliberate, not an
-    oversight. Overview's whole job is to show how a filtered population
-    splits ACROSS states, and the tiles double as the control that sets the
-    status filter. Honouring that filter here would mean clicking "Failed"
-    zeroes the other three tiles, i.e. the act of drilling in destroys the
-    context you drilled in from. So every other predicate applies and status
-    does not; the caller renders the active status as a selection instead.
-
-    Returns ``by_status`` (every state present, plus explicit zeros for the
-    four the queue writes, so the UI never has to invent a missing key),
-    ``by_target`` (per target format, split by state — this is what makes
-    "glb is fine, step is failing" visible at a glance), and
-    ``top_errors`` (the most common failure messages, so the usual next
-    question — *which* failure — is answered without opening a single row).
-    """
-    where, args = _audit_predicates(
-        user_sub=user_sub,
-        scope_kind=scope_kind,
-        scope_id=scope_id,
-        action=action,
-        target_format=target_format,
-        key_like=key_like,
-        since=since,
-        until=until,
-    )
-    clause = (" WHERE " + " AND ".join(where)) if where else ""
-
-    # One pass yields both breakdowns: summing over target gives the status
-    # totals, so the tiles and the per-target table can never disagree.
-    grid = await pool.fetch(
-        "SELECT status, target_format, count(*) AS n FROM audit_log" + clause + " GROUP BY status, target_format",
-        *args,
-    )
-
-    by_status: dict[str, int] = {"queued": 0, "running": 0, "done": 0, "error": 0}
-    by_target: dict[str, dict[str, int]] = {}
-    total = 0
-    for r in grid:
-        status = r["status"] or "unknown"
-        n = int(r["n"])
-        total += n
-        by_status[status] = by_status.get(status, 0) + n
-        tgt = r["target_format"] or "—"
-        bucket = by_target.setdefault(tgt, {})
-        bucket[status] = bucket.get(status, 0) + n
-
-    # CONGESTION. ``ts`` is the ENQUEUE time — insert_audit writes the row with
-    # status='queued' before a worker can see the job, and the completion update
-    # never rewrites it. So for a row still queued, ``now() - ts`` is exactly how
-    # long that job has been waiting, which is the question "are we backed up?"
-    #
-    # What this deliberately does NOT report is the wait of jobs that already
-    # ran. Nothing records when a worker picked a job up: the row carries
-    # enqueue time and processing duration, and the instant between them is
-    # simply not stored. A historical average would have to be invented, and an
-    # invented latency is worse than an absent one. Adding a ``started_at``
-    # column is the honest way to get it.
-    queue_row = await pool.fetchrow(
-        "SELECT count(*) AS n,"
-        " EXTRACT(EPOCH FROM max(now() - ts)) AS oldest_s,"
-        " EXTRACT(EPOCH FROM avg(now() - ts)) AS mean_s,"
-        " EXTRACT(EPOCH FROM percentile_cont(0.5) WITHIN GROUP (ORDER BY now() - ts)) AS median_s"
-        " FROM audit_log" + (clause + " AND " if clause else " WHERE ") + "status = 'queued'",
-        *args,
-    )
-
-    # HISTORICAL wait, now that started_at survives the hop that used to destroy
-    # it. Rows predating migration 027 have started_at NULL and are excluded
-    # rather than backfilled: giving them ts would report every historical job
-    # as having waited zero and flatten the very trend this measures.
-    served_row = await pool.fetchrow(
-        "SELECT count(*) AS n,"
-        " EXTRACT(EPOCH FROM avg(started_at - ts)) AS mean_s,"
-        " EXTRACT(EPOCH FROM percentile_cont(0.5) WITHIN GROUP (ORDER BY started_at - ts)) AS median_s,"
-        " EXTRACT(EPOCH FROM percentile_cont(0.95) WITHIN GROUP (ORDER BY started_at - ts)) AS p95_s,"
-        " EXTRACT(EPOCH FROM max(started_at - ts)) AS max_s"
-        " FROM audit_log" + (clause + " AND " if clause else " WHERE ") + "started_at IS NOT NULL",
-        *args,
-    )
-
-    err_where = list(where) + ["status = 'error'", "error IS NOT NULL"]
-    err_args = list(args)
-    err_args.append(min(max(reason_limit, 1), 50))
-    top_errors = await pool.fetch(
-        "SELECT error, count(*) AS n FROM audit_log WHERE "
-        + " AND ".join(err_where)
-        + f" GROUP BY error ORDER BY n DESC, error ASC LIMIT ${len(err_args)}",
-        *err_args,
-    )
-
-    def _sec(v):
-        return round(float(v), 1) if v is not None else None
-
-    return {
-        "total": total,
-        "by_status": by_status,
-        "congestion": {
-            "queued": int(queue_row["n"] or 0),
-            "running": by_status.get("running", 0),
-            # Ages of the jobs waiting RIGHT NOW, in seconds. None when the
-            # queue is empty — which is not the same as zero, and the UI says so.
-            "oldest_wait_s": _sec(queue_row["oldest_s"]),
-            "mean_wait_s": _sec(queue_row["mean_s"]),
-            "median_wait_s": _sec(queue_row["median_s"]),
-            # How long jobs that DID run waited before a worker took them.
-            # ``served`` is how many rows carry the measurement at all — rows
-            # from before migration 027 do not, and a median over three rows
-            # deserves less trust than one over three thousand, so the count
-            # travels with the numbers.
-            "served": int(served_row["n"] or 0),
-            "served_mean_wait_s": _sec(served_row["mean_s"]),
-            "served_median_wait_s": _sec(served_row["median_s"]),
-            "served_p95_wait_s": _sec(served_row["p95_s"]),
-            "served_max_wait_s": _sec(served_row["max_s"]),
-        },
-        "by_target": [
-            {"target": tgt, "counts": counts, "total": sum(counts.values())}
-            for tgt, counts in sorted(by_target.items(), key=lambda kv: -sum(kv[1].values()))
-        ],
-        "top_errors": [{"error": r["error"], "count": int(r["n"])} for r in top_errors],
-    }
-
-
-def _loads_jsonb(v):
-    """asyncpg may hand JSONB back as a str (no codec registered) or already
-    parsed. Normalize to a dict/list/None; never raise on a malformed row."""
-    if v is None or isinstance(v, (dict, list)):
-        return v
-    try:
-        return json.loads(v)
-    except (TypeError, ValueError):
-        return None
-
-
-async def upsert_worker_packages(pool: asyncpg.Pool, *, worker_image_tag: str, packages: list) -> None:
-    """Record a worker image's package manifest (idempotent per image tag)."""
-    await pool.execute(
-        """
-        INSERT INTO worker_packages (worker_image_tag, packages, captured_at)
-        VALUES ($1, $2, now())
-        ON CONFLICT (worker_image_tag) DO UPDATE
-          SET packages = EXCLUDED.packages, captured_at = now()
-        """,
-        worker_image_tag,
-        json.dumps(packages),
-    )
-
-
-async def get_worker_packages(pool: asyncpg.Pool, worker_image_tag: str) -> dict | None:
-    """The captured package manifest for a worker image tag (or None)."""
-    row = await pool.fetchrow(
-        "SELECT worker_image_tag, packages, captured_at FROM worker_packages WHERE worker_image_tag = $1",
-        worker_image_tag,
-    )
-    if row is None:
-        return None
-    return {
-        "worker_image_tag": row["worker_image_tag"],
-        "packages": _loads_jsonb(row["packages"]) or [],
-        "captured_at": row["captured_at"].isoformat() if row["captured_at"] else None,
-    }
-
-
-async def get_audit_by_job(pool: asyncpg.Pool, job_id: str) -> dict | None:
-    """Scope + source key for a queued job's audit row, newest first.
-
-    The worker patches its audit row by ``job_id`` and its error paths never
-    carry the scope, so failure capture reads it back from here rather than
-    threading a source key through every one of them. Narrow on purpose — this
-    runs on the failure path and only needs enough to address the blob.
-    """
-    row = await pool.fetchrow(
-        """
-        SELECT id, scope_kind, scope_id, key, action, status
-        FROM audit_log WHERE job_id = $1 ORDER BY id DESC LIMIT 1
-        """,
-        job_id,
-    )
-    if row is None:
-        return None
-    return {
-        "id": row["id"],
-        "scope_kind": row["scope_kind"],
-        "scope_id": row["scope_id"],
-        "key": row["key"],
-        "action": row["action"],
-        "status": row["status"],
-    }
-
-
-async def get_audit_by_id(pool: asyncpg.Pool, audit_id: int) -> dict | None:
-    """Fetch a single audit row by id. Used by the profile-download
-    endpoint to look up the blob key + scope without re-listing, and
-    by the local repro tooling to recover ``target_format`` + the
-    error context for a failed conversion."""
-    row = await pool.fetchrow(
-        """
-        SELECT id, ts, user_sub, scope_kind, scope_id, profile_key, log_key, failure_key, key,
-               action, target_format, status, error, traceback,
-               duration_ms, job_id, metrics_samples, audit_run_id,
-               issue_bot_status, issue_bot_synced_at, issue_bot_last_error
-        FROM audit_log WHERE id = $1
-        """,
-        audit_id,
-    )
-    if row is None:
-        return None
-    samples_raw = row["metrics_samples"]
-    # asyncpg returns JSONB as a Python str by default; parse defensively.
-    if isinstance(samples_raw, str):
-        try:
-            samples = json.loads(samples_raw)
-        except (ValueError, TypeError):
-            samples = None
-    else:
-        samples = samples_raw
-    return {
-        "id": row["id"],
-        "ts": row["ts"].isoformat() if row["ts"] is not None else None,
-        "user_sub": row["user_sub"],
-        "scope_kind": row["scope_kind"],
-        "scope_id": row["scope_id"],
-        "profile_key": row["profile_key"],
-        "log_key": row["log_key"],
-        "failure_key": row["failure_key"],
-        "key": row["key"],
-        "action": row["action"],
-        "target_format": row["target_format"],
-        "status": row["status"],
-        "error": row["error"],
-        "traceback": row["traceback"],
-        "duration_ms": row["duration_ms"],
-        "job_id": row["job_id"],
-        "metrics_samples": samples,
-        "audit_run_id": str(row["audit_run_id"]) if row["audit_run_id"] else None,
-        "issue_bot_status": row["issue_bot_status"],
-        "issue_bot_synced_at": (row["issue_bot_synced_at"].isoformat() if row["issue_bot_synced_at"] else None),
-        "issue_bot_last_error": row["issue_bot_last_error"],
-    }
-
-
-# ── App settings ─────────────────────────────────────────────────────
-
-
-async def get_setting(pool: asyncpg.Pool, key: str) -> str | None:
-    row = await pool.fetchrow("SELECT value FROM app_settings WHERE key = $1", key)
-    return row["value"] if row else None
-
-
-async def set_setting(pool: asyncpg.Pool, key: str, value: str, *, updated_by: str | None) -> None:
-    """Upsert a setting. ``value`` is a string — caller serializes
-    booleans / numbers as appropriate (we keep this small and avoid
-    type-tagging columns)."""
-    await pool.execute(
-        """
-        INSERT INTO app_settings (key, value, updated_by)
-        VALUES ($1, $2, $3)
-        ON CONFLICT (key) DO UPDATE SET
-            value = EXCLUDED.value,
-            updated_at = NOW(),
-            updated_by = EXCLUDED.updated_by
-        """,
-        key,
-        value,
-        updated_by,
-    )
-
-
-async def clear_audit_metrics(pool: asyncpg.Pool) -> dict:
-    """Null out the metrics columns on every audit row, returning
-    counts of rows touched and profile_keys that need blob cleanup.
-
-    The audit rows themselves are left intact — only the metrics
-    payload is wiped. Caller is responsible for deleting the actual
-    profile blobs from storage (we return the keys to make that
-    feasible without a second pass over the table).
-    """
-    profile_rows = await pool.fetch(
-        """
-        SELECT scope_kind, scope_id, profile_key
-        FROM audit_log
-        WHERE profile_key IS NOT NULL
-        """
-    )
-    profile_keys = [
-        {
-            "scope_kind": r["scope_kind"],
-            "scope_id": r["scope_id"],
-            "profile_key": r["profile_key"],
-        }
-        for r in profile_rows
-    ]
-    result = await pool.execute(
-        """
-        UPDATE audit_log
-        SET cpu_user_ms = NULL,
-            cpu_sys_ms = NULL,
-            peak_rss_kb = NULL,
-            read_bytes = NULL,
-            write_bytes = NULL,
-            profile_key = NULL
-        WHERE cpu_user_ms IS NOT NULL
-           OR cpu_sys_ms IS NOT NULL
-           OR peak_rss_kb IS NOT NULL
-           OR read_bytes IS NOT NULL
-           OR write_bytes IS NOT NULL
-           OR profile_key IS NOT NULL
-        """
-    )
-    # asyncpg returns "UPDATE N" — pull out the integer.
-    rows_cleared = 0
-    if isinstance(result, str) and result.startswith("UPDATE "):
-        try:
-            rows_cleared = int(result.split()[1])
-        except (IndexError, ValueError):
-            pass
-    return {"rows_cleared": rows_cleared, "profile_keys": profile_keys}
-
-
-async def list_all_projects(pool: asyncpg.Pool) -> list[dict]:
-    """Admin view: every project (including archived). Member count too."""
-    rows = await pool.fetch(
-        """
-        SELECT p.id, p.slug, p.name, p.created_at, p.archived_at,
-               COUNT(m.user_sub) AS member_count
-        FROM projects p
-        LEFT JOIN project_members m ON m.project_id = p.id
-        GROUP BY p.id
-        ORDER BY p.archived_at IS NOT NULL, p.name
-        """
-    )
-    return [
-        {
-            "id": str(r["id"]),
-            "slug": r["slug"],
-            "name": r["name"],
-            "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-            "archived_at": r["archived_at"].isoformat() if r["archived_at"] else None,
-            "member_count": int(r["member_count"]),
-        }
-        for r in rows
-    ]
-
-
-async def create_project(pool: asyncpg.Pool, slug: str, name: str) -> dict:
-    """Insert a project and return it. Slug is unique; conflicts → ValueError."""
-    try:
-        row = await pool.fetchrow(
-            """
-            INSERT INTO projects (slug, name) VALUES ($1, $2)
-            RETURNING id, slug, name, created_at, archived_at
-            """,
-            slug,
-            name,
-        )
-    except asyncpg.UniqueViolationError as exc:
-        raise ValueError(f"slug {slug!r} already exists") from exc
-    assert row is not None
-    return {
-        "id": str(row["id"]),
-        "slug": row["slug"],
-        "name": row["name"],
-        "created_at": row["created_at"].isoformat() if row["created_at"] else None,
-        "archived_at": row["archived_at"].isoformat() if row["archived_at"] else None,
-        "member_count": 0,
-    }
-
-
-async def archive_project(pool: asyncpg.Pool, project_id: str) -> bool:
-    """Soft-delete: stamp archived_at. Returns False when not found.
-
-    Soft delete preserves audit_log scope_id references and lets us
-    un-archive without orphaning. Hard-delete is intentionally not
-    exposed via the admin API.
-    """
-    row = await pool.fetchrow(
-        "UPDATE projects SET archived_at = NOW() WHERE id = $1 AND archived_at IS NULL RETURNING id",
-        project_id,
-    )
-    return row is not None
-
-
-async def list_project_members(pool: asyncpg.Pool, project_id: str) -> list[dict]:
-    rows = await pool.fetch(
-        """
-        SELECT m.user_sub, m.role, m.added_at,
-               u.email, u.display_name, u.last_seen_at
-        FROM project_members m
-        LEFT JOIN users u ON u.sub = m.user_sub
-        WHERE m.project_id = $1
-        ORDER BY u.display_name, m.user_sub
-        """,
-        project_id,
-    )
-    return [
-        {
-            "user_sub": r["user_sub"],
-            "role": r["role"],
-            "added_at": r["added_at"].isoformat() if r["added_at"] else None,
-            "email": r["email"],
-            "display_name": r["display_name"],
-            "last_seen_at": r["last_seen_at"].isoformat() if r["last_seen_at"] else None,
-        }
-        for r in rows
-    ]
-
-
-async def add_project_member(pool: asyncpg.Pool, project_id: str, user_sub: str, role: str = "member") -> bool:
-    """Idempotent membership add. Returns True on insert, False on duplicate.
-
-    Inserts a placeholder ``users`` row when the sub hasn't been seen
-    yet so the FK holds — the row gets enriched (email, display_name)
-    on the user's first authenticated request.
-    """
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            await conn.execute(
-                "INSERT INTO users (sub) VALUES ($1) ON CONFLICT (sub) DO NOTHING",
-                user_sub,
-            )
-            row = await conn.fetchrow(
-                """
-                INSERT INTO project_members (project_id, user_sub, role)
-                VALUES ($1, $2, $3)
-                ON CONFLICT (project_id, user_sub) DO NOTHING
-                RETURNING user_sub
-                """,
-                project_id,
-                user_sub,
-                role,
-            )
-    return row is not None
-
-
-async def remove_project_member(pool: asyncpg.Pool, project_id: str, user_sub: str) -> bool:
-    row = await pool.fetchrow(
-        "DELETE FROM project_members WHERE project_id = $1 AND user_sub = $2 RETURNING user_sub",
-        project_id,
-        user_sub,
-    )
-    return row is not None
-
-
-async def project_exists(pool: asyncpg.Pool, project_id: str) -> bool:
-    row = await pool.fetchrow("SELECT 1 FROM projects WHERE id = $1", project_id)
-    return row is not None
-
-
-async def project_id_from_slug(pool: asyncpg.Pool, slug: str) -> str | None:
-    """Resolve a project slug to its UUID. Returns None if no match.
-
-    Slugs are URL-safe and stable across renames of the human-readable
-    name; UUIDs are the FK-stable id. The scope URL parser uses this to
-    accept ``project:<slug>`` as a friendlier alternative to
-    ``project:<uuid>``.
-    """
-    row = await pool.fetchrow(
-        "SELECT id FROM projects WHERE slug = $1 AND archived_at IS NULL",
-        slug,
-    )
-    return str(row["id"]) if row else None
-
-
-# ── Corpora (M3 admin audit panel) ──────────────────────────────────
-
-
-def _corpus_row(r) -> dict:
-    return {
-        "id": str(r["id"]),
-        "slug": r["slug"],
-        "name": r["name"],
-        "description": r["description"],
-        "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-        "created_by": r["created_by"],
-        "archived_at": (r["archived_at"].isoformat() if r["archived_at"] else None),
-    }
-
-
-async def list_corpora(
-    pool: asyncpg.Pool,
-    *,
-    include_archived: bool = False,
-) -> list[dict]:
-    """Newest-first corpora list. Archived rows hidden by default; the
-    admin panel never needs them and the audit-form picker would
-    confuse the operator by listing slugs that storage will reject."""
-    where = "" if include_archived else " WHERE archived_at IS NULL"
-    rows = await pool.fetch(
-        f"""
-        SELECT id, slug, name, description, created_at,
-               created_by, archived_at
-        FROM corpora
-        {where}
-        ORDER BY created_at DESC
-        """
-    )
-    return [_corpus_row(r) for r in rows]
-
-
-async def create_corpus(
-    pool: asyncpg.Pool,
-    *,
-    slug: str,
-    name: str,
-    description: str | None = None,
-    created_by: str | None = None,
-) -> dict:
-    """Insert one new corpus. ``slug`` is the human-readable id used
-    on the wire (``corpus:<slug>``); enforce uniqueness against live
-    rows via the partial unique index. Archived rows holding the
-    same slug are no obstacle."""
-    row = await pool.fetchrow(
-        """
-        INSERT INTO corpora (slug, name, description, created_by)
-        VALUES ($1, $2, $3, $4)
-        RETURNING id, slug, name, description, created_at,
-                  created_by, archived_at
-        """,
-        slug,
-        name,
-        description,
-        created_by,
-    )
-    return _corpus_row(row)
-
-
-async def update_corpus(
-    pool: asyncpg.Pool,
-    slug: str,
-    *,
-    name: str,
-    description: str | None,
-) -> dict | None:
-    """Update a live corpus's display name / description. The slug is
-    immutable — it's baked into the storage prefix (``corpus/<slug>/``)
-    and scope URLs, so renaming it would orphan the bucket bytes.
-    Returns the updated row, or None when the slug isn't live."""
-    row = await pool.fetchrow(
-        """
-        UPDATE corpora
-        SET name = $2, description = $3
-        WHERE slug = $1 AND archived_at IS NULL
-        RETURNING id, slug, name, description, created_at,
-                  created_by, archived_at
-        """,
-        slug,
-        name,
-        description,
-    )
-    return _corpus_row(row) if row else None
-
-
-async def get_corpus_by_slug(
-    pool: asyncpg.Pool,
-    slug: str,
-) -> dict | None:
-    """Look up one corpus by its public slug. Returns the live row
-    only; archived corpora are treated as absent."""
-    row = await pool.fetchrow(
-        """
-        SELECT id, slug, name, description, created_at,
-               created_by, archived_at
-        FROM corpora
-        WHERE slug = $1 AND archived_at IS NULL
-        """,
-        slug,
-    )
-    return _corpus_row(row) if row else None
-
-
-async def archive_corpus(pool: asyncpg.Pool, slug: str) -> bool:
-    """Soft-delete a corpus. Storage bytes stay in the bucket — the
-    operator wipes those separately if disk pressure matters. The
-    slug becomes available for re-use immediately because the
-    uniqueness index is partial-on-live."""
-    result = await pool.execute(
-        """
-        UPDATE corpora
-        SET archived_at = NOW()
-        WHERE slug = $1 AND archived_at IS NULL
-        """,
-        slug,
-    )
-    return result.endswith(" 1")
-
-
-# ── Audit runs (M1 admin audit panel) ──────────────────────────────
-
-
-def _audit_run_row(r) -> dict:
-    """Project an audit_runs row to its JSON-ready dict shape.
-
-    Includes the M5 issue-bot fields when the underlying row has
-    them (it always does post-migration 009, but the helper tolerates
-    rows from a SELECT that omits those columns by falling back to
-    ``None``)."""
-
-    def _opt(col: str):
-        try:
-            return r[col]
-        except (KeyError, TypeError):
-            return None
-
-    issue_bot_synced_at = _opt("issue_bot_synced_at")
-    parent_run_id = _opt("parent_run_id")
-    av_dispatched = _opt("auto_validate_dispatched_at")
-    return {
-        "id": str(r["id"]),
-        "scope": r["scope"],
-        "worker_pool": r["worker_pool"],
-        "trigger": r["trigger"],
-        "started_at": r["started_at"].isoformat() if r["started_at"] else None,
-        "finished_at": r["finished_at"].isoformat() if r["finished_at"] else None,
-        "status": r["status"],
-        "note": r["note"],
-        "total": r["total"],
-        # Parity cells reserved in ``total`` for the auto-validation pass but
-        # not yet enqueued (0 once the pass dispatches, or for non-auto runs).
-        "validate_total": _opt("validate_total") or 0,
-        "ok": r["ok"],
-        "failed": r["failed"],
-        "skipped": r["skipped"],
-        "created_by": r["created_by"],
-        "seq": _opt("seq"),
-        # Idle time excluded from the run's active duration (gap before a later
-        # validation pass). The UI shows (finished_at - started_at) - idle_ms.
-        "idle_ms": _opt("idle_ms") or 0,
-        # Sum of every cell's own duration_ms — the run's active compute time,
-        # independent of wall clock (which parallel workers compress and a
-        # single-cell re-run inflates with idle). Recomputes whenever a cell's
-        # row changes, so a re-run's new timing is reflected immediately.
-        "cells_duration_ms": _opt("cells_duration_ms") or 0,
-        "force_rebuild": _opt("force_rebuild") or False,
-        "auto_validate": _opt("auto_validate") or False,
-        "parent_run_id": str(parent_run_id) if parent_run_id else None,
-        "auto_validate_dispatched_at": (av_dispatched.isoformat() if av_dispatched else None),
-        "issue_bot_status": _opt("issue_bot_status"),
-        "issue_bot_last_error": _opt("issue_bot_last_error"),
-        "issue_bot_synced_at": (issue_bot_synced_at.isoformat() if issue_bot_synced_at else None),
-    }
-
-
-async def create_audit_run(
-    pool: asyncpg.Pool,
-    *,
-    scope: str,
-    worker_pool: str | None,
-    trigger: str = "manual",
-    note: str | None = None,
-    created_by: str | None = None,
-    force_rebuild: bool = False,
-    auto_validate: bool = False,
-    parent_run_id: str | None = None,
-) -> dict:
-    """Open a new audit_runs row in ``status='running'``. Returns the
-    fresh row (including its server-generated UUID + started_at) so
-    the dispatcher can stamp the jobs it enqueues. ``total`` starts
-    at 0 — :func:`set_audit_run_total` finalises it once dispatch
-    knows how many jobs landed.
-
-    ``force_rebuild`` (M7+) bypasses the dispatcher's cached-cell
-    short-circuit so every viable cell actually re-converts.
-    Persisted on the row so the admin UI can show "this was a
-    force-rebuild" badge.
-
-    ``auto_validate`` asks the finished-run poller to fire a follow-up
-    ``validate_only`` parity run once this run finishes. ``parent_run_id``
-    links a derived run (the validation child, or a re-dispatched copy)
-    back to its source.
-    """
-    row = await pool.fetchrow(
-        """
-        INSERT INTO audit_runs
-            (scope, worker_pool, trigger, note, created_by, force_rebuild,
-             auto_validate, parent_run_id)
-        VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
-        RETURNING *
-        """,
-        scope,
-        worker_pool,
-        trigger,
-        note,
-        created_by,
-        force_rebuild,
-        auto_validate,
-        parent_run_id,
-    )
-    return _audit_run_row(row)
-
-
-async def set_audit_run_total(
-    pool: asyncpg.Pool,
-    run_id: str,
-    total: int,
-    *,
-    validate_total: int = 0,
-) -> None:
-    """Set the dispatched-job count after enqueue completes. If
-    ``total`` is 0 (no jobs to run — empty scope or no viable cells),
-    the run is marked ``finished`` immediately so the UI doesn't show
-    a perpetually-running row.
-
-    ``validate_total`` reserves that many of the ``total`` cells for the
-    auto-validation parity pass, which is dispatched later (once the
-    conversion cells have landed) — so the run's total is complete from
-    the start. The reservation is consumed by
-    :func:`consume_audit_run_validation_reserve`."""
-    await pool.execute(
-        """
-        UPDATE audit_runs
-        SET total = $2,
-            validate_total = $3,
-            status = CASE WHEN $2 = 0 THEN 'finished' ELSE status END,
-            finished_at = CASE WHEN $2 = 0 THEN NOW() ELSE finished_at END
-        WHERE id = $1
-        """,
-        run_id,
-        total,
-        validate_total,
-    )
-
-
-async def extend_audit_run_total(pool: asyncpg.Pool, run_id: str, delta: int) -> None:
-    """Grow a finished run's ``total`` by ``delta`` and reopen it to
-    ``running`` (clearing ``finished_at``). Used to append cells — an
-    auto-validation parity pass — to an existing run instead of spawning a
-    separate one: once the appended cells land, the normal counter-bump
-    re-finishes the same run. Aborted runs are left aborted. Caller guards
-    ``delta > 0``."""
-    await pool.execute(
-        """
-        UPDATE audit_runs
-        SET total = total + $2,
-            -- Fold the idle gap (since this run last finished) into idle_ms so
-            -- the displayed duration stays block-active time, not wall clock.
-            -- RHS reads the pre-update finished_at / status.
-            idle_ms = idle_ms + CASE
-                WHEN status <> 'aborted' AND finished_at IS NOT NULL
-                THEN GREATEST(0, (EXTRACT(EPOCH FROM (NOW() - finished_at)) * 1000)::bigint)
-                ELSE 0 END,
-            status = CASE WHEN status = 'aborted' THEN 'aborted' ELSE 'running' END,
-            finished_at = CASE WHEN status = 'aborted' THEN finished_at ELSE NULL END
-        WHERE id = $1
-        """,
-        run_id,
-        delta,
-    )
-
-
-async def consume_audit_run_validation_reserve(
-    pool: asyncpg.Pool,
-    run_id: str,
-    actual: int,
-) -> None:
-    """Swap a run's reserved validation-cell count for the ``actual``
-    number of parity cells enumerated at dispatch time.
-
-    ``total`` was set upfront to conversions + reservation; here it is
-    corrected by (actual - validate_total) — normally zero, non-zero only
-    when the scope's files changed between the two enumerations — and the
-    reservation is cleared. If the corrected total is already met (e.g.
-    ``actual`` is 0 because the parity enumeration failed or the sources
-    vanished), the run finishes here since no counter-bump will arrive to
-    do it. Aborted runs only get their bookkeeping corrected."""
-    await pool.execute(
-        """
-        UPDATE audit_runs
-        SET total = total - validate_total + $2,
-            validate_total = 0,
-            finished_at = CASE
-                WHEN status = 'aborted' THEN finished_at
-                WHEN ok + failed + skipped >= total - validate_total + $2
-                    THEN COALESCE(finished_at, NOW())
-                ELSE finished_at
-            END,
-            status = CASE
-                WHEN status = 'aborted' THEN 'aborted'
-                WHEN ok + failed + skipped >= total - validate_total + $2
-                    THEN 'finished'
-                ELSE status
-            END
-        WHERE id = $1
-        """,
-        run_id,
-        actual,
-    )
-
-
-async def delete_audit_run(pool: asyncpg.Pool, run_id: str) -> tuple[bool, list[str]]:
-    """Delete an audit run and its audit_log rows. ``audit_parity`` rows cascade
-    on the run delete; ``audit_log`` is ON DELETE SET NULL, so we remove those
-    rows explicitly rather than orphaning them as run-less audit entries.
-
-    Returns ``(deleted, queued_job_ids)``: ``deleted`` is True if a run row was
-    removed; ``queued_job_ids`` are the still-queued cells' job_ids captured BEFORE
-    the delete, so the caller can purge their JetStream messages (deleting the rows
-    would otherwise make a cancelled cell invisible to the worker's cancel check)."""
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            queued = await conn.fetch(
-                "SELECT job_id FROM audit_log WHERE audit_run_id = $1 AND status IN ('queued', 'running')",
-                run_id,
-            )
-            await conn.execute("DELETE FROM audit_log WHERE audit_run_id = $1", run_id)
-            res = await conn.execute("DELETE FROM audit_runs WHERE id = $1", run_id)
-    # asyncpg returns a command tag like "DELETE 1".
-    return res.split()[-1] != "0", [r["job_id"] for r in queued if r["job_id"]]
-
-
-async def list_audit_runs(
-    pool: asyncpg.Pool,
-    *,
-    limit: int = 50,
-    before_started_at: str | None = None,
-) -> list[dict]:
-    """Reverse-chronological audit_runs scan. Keyset paginated on
-    ``started_at`` so new runs landing between requests don't shift
-    the page boundary the way an offset would."""
-    args: list = []
-    where = ""
-    if before_started_at:
-        args.append(before_started_at)
-        where = f" WHERE started_at < ${len(args)}::timestamptz"
-    args.append(min(max(limit, 1), 200))
-    rows = await pool.fetch(
-        f"""
-        SELECT id, seq, scope, worker_pool, trigger, started_at, finished_at,
-               status, note, total, validate_total, ok, failed, skipped, created_by, idle_ms,
-               force_rebuild, auto_validate, parent_run_id, auto_validate_dispatched_at,
-               issue_bot_status, issue_bot_last_error, issue_bot_synced_at,
-               (SELECT COALESCE(SUM(duration_ms), 0)::bigint FROM audit_log
-                WHERE audit_run_id = audit_runs.id) AS cells_duration_ms
-        FROM audit_runs
-        {where}
-        ORDER BY started_at DESC
-        LIMIT ${len(args)}
-        """,
-        *args,
-    )
-    return [_audit_run_row(r) for r in rows]
-
-
-async def get_audit_run(pool: asyncpg.Pool, run_id: str) -> dict | None:
-    row = await pool.fetchrow(
-        """
-        SELECT id, seq, scope, worker_pool, trigger, started_at, finished_at,
-               status, note, total, validate_total, ok, failed, skipped, created_by, idle_ms,
-               force_rebuild, auto_validate, parent_run_id, auto_validate_dispatched_at,
-               issue_bot_status, issue_bot_last_error, issue_bot_synced_at,
-               (SELECT COALESCE(SUM(duration_ms), 0)::bigint FROM audit_log
-                WHERE audit_run_id = audit_runs.id) AS cells_duration_ms
-        FROM audit_runs WHERE id = $1
-        """,
-        run_id,
-    )
-    return _audit_run_row(row) if row else None
-
-
-async def list_audit_run_jobs(
-    pool: asyncpg.Pool,
-    run_id: str,
-) -> list[dict]:
-    """Every audit_log row tied to one audit_run. Returned in
-    insert order (ascending id) so the per-run grid in the admin
-    panel can render rows in the deterministic order the
-    dispatcher emitted them."""
-    rows = await pool.fetch(
-        """
-        SELECT id, ts, key, target_format, status, error,
-               duration_ms, cpu_user_ms, cpu_sys_ms, peak_rss_kb,
-               read_bytes, write_bytes, job_id, worker_image_tag, convert_meta
-        FROM audit_log
-        WHERE audit_run_id = $1
-        ORDER BY id ASC
-        """,
-        run_id,
-    )
-    return [
-        {
-            "id": r["id"],
-            "ts": r["ts"].isoformat() if r["ts"] else None,
-            "key": r["key"],
-            "target_format": r["target_format"],
-            "status": r["status"],
-            "error": r["error"],
-            "duration_ms": r["duration_ms"],
-            "cpu_user_ms": r["cpu_user_ms"],
-            "cpu_sys_ms": r["cpu_sys_ms"],
-            "peak_rss_kb": r["peak_rss_kb"],
-            "read_bytes": r["read_bytes"],
-            "write_bytes": r["write_bytes"],
-            "job_id": r["job_id"],
-            "worker_image_tag": r["worker_image_tag"],
-            "convert_meta": _loads_jsonb(r["convert_meta"]),
-        }
-        for r in rows
-    ]
-
-
-async def insert_audit_parity(
-    pool: asyncpg.Pool,
-    *,
-    job_id: str | None,
-    source_key: str,
-    baseline: int,
-    counts: dict[str, int],
-    consistent: bool,
-    mismatches: dict[str, int],
-    errors: dict[str, str],
-) -> None:
-    """Record one cross-format parity result. The owning ``audit_run_id`` is
-    resolved from the parity job's audit_log row (the worker doesn't carry it)."""
-    await pool.execute(
-        """
-        INSERT INTO audit_parity
-            (audit_run_id, job_id, source_key, baseline, counts, consistent, mismatches, errors)
-        VALUES (
-            (SELECT audit_run_id FROM audit_log WHERE job_id = $1 ORDER BY id DESC LIMIT 1),
-            $1, $2, $3, $4::jsonb, $5, $6::jsonb, $7::jsonb
-        )
-        """,
-        job_id,
-        source_key,
-        baseline,
-        json.dumps(counts),
-        consistent,
-        json.dumps(mismatches),
-        json.dumps(errors),
-    )
-
-
-async def list_audit_run_parity(pool: asyncpg.Pool, run_id: str) -> list[dict]:
-    """Every parity result for one audit_run, newest first."""
-    rows = await pool.fetch(
-        """
-        SELECT id, ts, job_id, source_key, baseline, counts, consistent, mismatches, errors
-        FROM audit_parity
-        WHERE audit_run_id = $1
-        ORDER BY id DESC
-        """,
-        run_id,
-    )
-    return [
-        {
-            "id": r["id"],
-            "ts": r["ts"].isoformat() if r["ts"] else None,
-            "job_id": r["job_id"],
-            "source_key": r["source_key"],
-            "baseline": r["baseline"],
-            "counts": json.loads(r["counts"]) if isinstance(r["counts"], str) else r["counts"],
-            "consistent": r["consistent"],
-            "mismatches": json.loads(r["mismatches"]) if isinstance(r["mismatches"], str) else r["mismatches"],
-            "errors": json.loads(r["errors"]) if isinstance(r["errors"], str) else r["errors"],
-        }
-        for r in rows
-    ]
-
-
-async def audit_run_exists_for_key(
-    pool: asyncpg.Pool,
-    scope: str,
-    worker_pool: str | None,
-) -> bool:
-    """Concurrent-fire guard for the scheduler tick (M4). Returns True
-    when an ``audit_runs`` row with the same (scope, worker_pool) is
-    still ``status='running'``. ``worker_pool=None`` is matched against
-    SQL NULL explicitly so the "any pool" tag doesn't collide with a
-    schedule that pins a specific pool."""
-    if worker_pool is None:
-        row = await pool.fetchrow(
-            """
-            SELECT 1 FROM audit_runs
-            WHERE status = 'running'
-              AND scope = $1
-              AND worker_pool IS NULL
-            LIMIT 1
-            """,
-            scope,
-        )
-    else:
-        row = await pool.fetchrow(
-            """
-            SELECT 1 FROM audit_runs
-            WHERE status = 'running'
-              AND scope = $1
-              AND worker_pool = $2
-            LIMIT 1
-            """,
-            scope,
-            worker_pool,
-        )
-    return row is not None
-
-
-# ── Audit schedules (M4 admin audit panel) ─────────────────────────
-
-
-def _audit_schedule_row(r) -> dict:
-    """Project an audit_schedules row to its JSON-ready dict shape."""
-    return {
-        "id": str(r["id"]),
-        "name": r["name"],
-        "cron_expr": r["cron_expr"],
-        "scope": r["scope"],
-        "worker_pool": r["worker_pool"],
-        "enabled": r["enabled"],
-        "last_fired_at": (r["last_fired_at"].isoformat() if r["last_fired_at"] else None),
-        "next_fire_at": (r["next_fire_at"].isoformat() if r["next_fire_at"] else None),
-        "last_skipped_reason": r["last_skipped_reason"],
-        "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-        "created_by": r["created_by"],
-        "archived_at": (r["archived_at"].isoformat() if r["archived_at"] else None),
-    }
-
-
-_SCHEDULE_COLS = (
-    "id, name, cron_expr, scope, worker_pool, enabled, "
-    "last_fired_at, next_fire_at, last_skipped_reason, "
-    "created_at, created_by, archived_at"
-)
-
-
-async def list_audit_schedules(
-    pool: asyncpg.Pool,
-    *,
-    include_archived: bool = False,
-) -> list[dict]:
-    """Newest-first audit_schedules listing. Archived rows hidden by
-    default — the admin panel's primary view should only see live
-    schedules."""
-    where = "" if include_archived else " WHERE archived_at IS NULL"
-    rows = await pool.fetch(f"SELECT {_SCHEDULE_COLS} FROM audit_schedules{where} " "ORDER BY created_at DESC")
-    return [_audit_schedule_row(r) for r in rows]
-
-
-async def get_audit_schedule(
-    pool: asyncpg.Pool,
-    schedule_id: str,
-) -> dict | None:
-    row = await pool.fetchrow(
-        f"SELECT {_SCHEDULE_COLS} FROM audit_schedules WHERE id = $1",
-        schedule_id,
-    )
-    return _audit_schedule_row(row) if row else None
-
-
-async def create_audit_schedule(
-    pool: asyncpg.Pool,
-    *,
-    name: str,
-    cron_expr: str,
-    scope: str,
-    worker_pool: str | None,
-    next_fire_at,
-    enabled: bool = True,
-    created_by: str | None = None,
-) -> dict:
-    """Insert one new audit_schedule. ``next_fire_at`` is pre-computed
-    by the caller via croniter — keeping the cron parsing in the
-    route handler means the DB layer stays library-free and the 400
-    path for malformed expressions is sharper."""
-    row = await pool.fetchrow(
-        f"""
-        INSERT INTO audit_schedules
-            (name, cron_expr, scope, worker_pool, enabled, next_fire_at, created_by)
-        VALUES ($1, $2, $3, $4, $5, $6, $7)
-        RETURNING {_SCHEDULE_COLS}
-        """,
-        name,
-        cron_expr,
-        scope,
-        worker_pool,
-        enabled,
-        next_fire_at,
-        created_by,
-    )
-    return _audit_schedule_row(row)
-
-
-async def update_audit_schedule(
-    pool: asyncpg.Pool,
-    schedule_id: str,
-    *,
-    name: str | None = None,
-    cron_expr: str | None = None,
-    scope: str | None = None,
-    worker_pool: str | None = None,
-    worker_pool_set: bool = False,
-    enabled: bool | None = None,
-    next_fire_at=None,
-    next_fire_at_set: bool = False,
-) -> dict | None:
-    """Partial update. Only fields with non-None values (and the
-    ``*_set`` toggles for nullable columns where ``None`` is a
-    meaningful payload value) are written; the rest are left at
-    whatever the row already holds."""
-    sets: list[str] = []
-    args: list = []
-    if name is not None:
-        args.append(name)
-        sets.append(f"name = ${len(args)}")
-    if cron_expr is not None:
-        args.append(cron_expr)
-        sets.append(f"cron_expr = ${len(args)}")
-    if scope is not None:
-        args.append(scope)
-        sets.append(f"scope = ${len(args)}")
-    if worker_pool_set:
-        args.append(worker_pool)
-        sets.append(f"worker_pool = ${len(args)}")
-    if enabled is not None:
-        args.append(enabled)
-        sets.append(f"enabled = ${len(args)}")
-    if next_fire_at_set:
-        args.append(next_fire_at)
-        sets.append(f"next_fire_at = ${len(args)}")
-    if not sets:
-        return await get_audit_schedule(pool, schedule_id)
-    args.append(schedule_id)
-    row = await pool.fetchrow(
-        f"""
-        UPDATE audit_schedules
-        SET {', '.join(sets)}
-        WHERE id = ${len(args)} AND archived_at IS NULL
-        RETURNING {_SCHEDULE_COLS}
-        """,
-        *args,
-    )
-    return _audit_schedule_row(row) if row else None
-
-
-async def archive_audit_schedule(
-    pool: asyncpg.Pool,
-    schedule_id: str,
-) -> bool:
-    """Soft-delete one schedule. The tick query filters on
-    ``archived_at IS NULL`` so the row stops firing immediately."""
-    result = await pool.execute(
-        "UPDATE audit_schedules SET archived_at = NOW() " "WHERE id = $1 AND archived_at IS NULL",
-        schedule_id,
-    )
-    return result.endswith(" 1")
-
-
-async def claim_due_audit_schedule(
-    pool: asyncpg.Pool,
-    *,
-    now,
-    next_fire_at,
-):
-    """Atomically claim ONE due schedule.
-
-    Returns the claimed row (as a dict) or ``None`` if nothing was
-    due. The UPDATE pins ``last_fired_at`` to the provided ``now``,
-    advances ``next_fire_at`` to the supplied value (computed by the
-    caller via croniter from ``now``), and clears
-    ``last_skipped_reason`` — the caller can re-set the reason via
-    :func:`set_audit_schedule_skip_reason` if dispatch ends up not
-    firing.
-
-    Uses ``FOR UPDATE SKIP LOCKED`` so multiple API replicas can tick
-    in parallel without ever firing the same schedule twice.
-    """
-    row = await pool.fetchrow(
-        f"""
-        UPDATE audit_schedules
-        SET last_fired_at = $1,
-            next_fire_at = $2,
-            last_skipped_reason = NULL
-        WHERE id = (
-            SELECT id FROM audit_schedules
-            WHERE enabled
-              AND archived_at IS NULL
-              AND next_fire_at IS NOT NULL
-              AND next_fire_at <= $1
-            ORDER BY next_fire_at ASC
-            LIMIT 1
-            FOR UPDATE SKIP LOCKED
-        )
-        RETURNING {_SCHEDULE_COLS}
-        """,
-        now,
-        next_fire_at,
-    )
-    return _audit_schedule_row(row) if row else None
-
-
-async def set_audit_schedule_skip_reason(
-    pool: asyncpg.Pool,
-    schedule_id: str,
-    reason: str,
-) -> None:
-    """Record why a tick decided not to dispatch (concurrent-fire
-    guard, scope resolution failure, etc.). Stamped after the atomic
-    claim so the operator sees the latest reason without unwinding
-    the ``last_fired_at`` bump — the schedule still advances to the
-    next slot rather than re-trying the missed one."""
-    await pool.execute(
-        "UPDATE audit_schedules SET last_skipped_reason = $2 WHERE id = $1",
-        schedule_id,
-        reason,
-    )
-
-
-# ── Audit issue-bot (M5 admin audit panel) ─────────────────────────
-
-
-async def claim_audit_run_for_issue_bot(
-    pool: asyncpg.Pool,
-):
-    """Atomically claim the oldest finished audit_run that hasn't
-    been issue-synced yet. Sets ``issue_bot_status='syncing'`` so
-    other replicas / retries skip it.
-
-    Returns the claimed row (dict shaped like the public audit_run
-    projection) or ``None`` if nothing is pending. The caller is
-    expected to call :func:`mark_audit_run_issue_bot` with a
-    terminal state once the sync completes (or fails).
-    """
-    row = await pool.fetchrow(
-        """
-        UPDATE audit_runs
-        SET issue_bot_status = 'syncing'
-        WHERE id = (
-            SELECT id FROM audit_runs
-            WHERE status = 'finished'
-              AND issue_bot_status IS NULL
-            ORDER BY finished_at ASC NULLS LAST
-            LIMIT 1
-            FOR UPDATE SKIP LOCKED
-        )
-        RETURNING id, scope, worker_pool, trigger, started_at, finished_at,
-                  status, note, total, ok, failed, skipped, created_by,
-                  force_rebuild,
-                  issue_bot_status, issue_bot_last_error, issue_bot_synced_at
-        """,
-    )
-    return _audit_run_row(row) if row else None
-
-
-async def claim_audit_run_for_auto_validate(pool: asyncpg.Pool):
-    """Atomically claim the oldest ``auto_validate`` run that is ready for its
-    validation pass. Two shapes qualify:
-
-    * a run whose parity cells were reserved upfront (``validate_total > 0``)
-      and whose conversion cells have all landed — it is still ``running``
-      because the reserved cells keep the counter-bump finish check from
-      tripping;
-    * a pre-reservation run (``validate_total = 0``) that already finished —
-      its validation extends the total, the legacy behaviour.
-
-    Stamps ``auto_validate_dispatched_at`` in the same statement so a
-    concurrent tick / replica can't double-fire. Returns the claimed run row
-    or ``None``."""
-    row = await pool.fetchrow(
-        """
-        UPDATE audit_runs
-        SET auto_validate_dispatched_at = NOW()
-        WHERE id = (
-            SELECT id FROM audit_runs
-            WHERE auto_validate = TRUE
-              AND auto_validate_dispatched_at IS NULL
-              AND (
-                    (validate_total > 0 AND status = 'running'
-                        AND ok + failed + skipped >= total - validate_total)
-                 OR (validate_total = 0 AND status = 'finished')
-              )
-            ORDER BY started_at ASC NULLS LAST
-            LIMIT 1
-            FOR UPDATE SKIP LOCKED
-        )
-        RETURNING id, scope, worker_pool, trigger, started_at, finished_at,
-                  status, note, total, validate_total, ok, failed, skipped,
-                  created_by, force_rebuild, auto_validate, parent_run_id,
-                  issue_bot_status, issue_bot_last_error, issue_bot_synced_at
-        """,
-    )
-    return _audit_run_row(row) if row else None
-
-
-async def claim_run_for_validation(pool: asyncpg.Pool, run_id: str):
-    """Atomically mark one *specific* finished run as having its validation
-    pass dispatched (the manual 'Validate' button). Returns the run if newly
-    claimed, or ``None`` when it's already been validated, isn't finished, or
-    doesn't exist — so a validation is appended at most once per run, whether
-    via the auto-validate toggle or the button."""
-    row = await pool.fetchrow(
-        """
-        UPDATE audit_runs
-        SET auto_validate_dispatched_at = NOW()
-        WHERE id = $1
-          AND status = 'finished'
-          AND auto_validate_dispatched_at IS NULL
-        RETURNING id, scope, worker_pool, trigger, started_at, finished_at,
-                  status, note, total, ok, failed, skipped, created_by,
-                  force_rebuild, auto_validate, parent_run_id, auto_validate_dispatched_at,
-                  issue_bot_status, issue_bot_last_error, issue_bot_synced_at
-        """,
-        run_id,
-    )
-    return _audit_run_row(row) if row else None
-
-
-async def audit_log_history_for_cell(
-    pool: asyncpg.Pool,
-    key: str,
-    target_format: str,
-    *,
-    limit: int = 50,
-) -> list[dict]:
-    """Historic ``audit_log`` results for one ``(source key, target_format)``
-    cell across every run — newest first. Powers the per-cell 'show history'
-    table so an operator can spot run-to-run regressions for one conversion.
-
-    Only rows tied to an audit run (``audit_run_id IS NOT NULL``) are returned:
-    history is a cross-run comparison, so on-demand / viewer-triggered
-    re-conversions (e.g. a GLB re-tessellation on file open, ``audit_run_id``
-    NULL) are excluded — otherwise a heavily-viewed source floods its glb cell
-    with standalone rows that push the actual per-run results past ``limit``."""
-    rows = await pool.fetch(
-        """
-        SELECT id, ts, status, error, duration_ms, peak_rss_kb,
-               worker_image_tag, audit_run_id
-        FROM audit_log
-        WHERE key = $1 AND target_format = $2 AND audit_run_id IS NOT NULL
-        ORDER BY id DESC
-        LIMIT $3
-        """,
-        key,
-        target_format,
-        min(max(limit, 1), 500),
-    )
-    return [
-        {
-            "id": r["id"],
-            "ts": r["ts"].isoformat() if r["ts"] else None,
-            "status": r["status"],
-            "error": r["error"],
-            "duration_ms": r["duration_ms"],
-            "peak_rss_kb": r["peak_rss_kb"],
-            "worker_image_tag": r["worker_image_tag"],
-            "audit_run_id": str(r["audit_run_id"]) if r["audit_run_id"] else None,
-        }
-        for r in rows
-    ]
-
-
-async def mark_audit_run_issue_bot(
-    pool: asyncpg.Pool,
-    run_id: str,
-    *,
-    status: str,
-    error: str | None = None,
-) -> None:
-    """Stamp a terminal issue-bot status on the run. ``status`` is
-    'done' (issues synced), 'skipped' (no failures to sync / bot
-    disabled), or 'failed' (raised mid-sync — ``error`` carries the
-    summary). All three set ``issue_bot_synced_at`` to NOW() so the
-    UI can show how recently the bot ran."""
-    await pool.execute(
-        """
-        UPDATE audit_runs
-        SET issue_bot_status = $2,
-            issue_bot_last_error = $3,
-            issue_bot_synced_at = NOW()
-        WHERE id = $1
-        """,
-        run_id,
-        status,
-        error,
-    )
-
-
-async def reset_audit_run_issue_bot(
-    pool: asyncpg.Pool,
-    run_id: str,
-) -> bool:
-    """Clear the issue-bot status so the next tick picks the run up
-    again. Used by the admin "retry sync" button. Returns True on a
-    real reset, False when the run wasn't found or wasn't finished
-    (in which case retrying makes no sense)."""
-    result = await pool.execute(
-        """
-        UPDATE audit_runs
-        SET issue_bot_status = NULL,
-            issue_bot_last_error = NULL,
-            issue_bot_synced_at = NULL
-        WHERE id = $1 AND status = 'finished'
-        """,
-        run_id,
-    )
-    return result.endswith(" 1")
-
-
-# ── Single-conversion issue-bot (M5b) ──────────────────────────────
-
-
-async def claim_failed_conversion_for_issue_bot(
-    pool: asyncpg.Pool,
-) -> dict | None:
-    """Atomically claim the oldest failed user-driven conversion
-    that hasn't been synced yet.
-
-    Restricted to ``audit_run_id IS NULL`` rows so audit-sweep
-    failures keep going through the parent run's bot pass — that
-    path batches all of a run's failures into one sync, which is
-    much friendlier on the forge API than firing N times.
-    """
-    row = await pool.fetchrow(
-        """
-        UPDATE audit_log
-        SET issue_bot_status = 'syncing'
-        WHERE id = (
-            SELECT id FROM audit_log
-            WHERE status IN ('error', 'failed')
-              AND audit_run_id IS NULL
-              AND issue_bot_status IS NULL
-            ORDER BY id ASC
-            LIMIT 1
-            FOR UPDATE SKIP LOCKED
-        )
-        RETURNING id, ts, user_sub, scope_kind, scope_id, action,
-                  key, target_format, status, error, traceback
-        """,
-    )
-    if row is None:
-        return None
-    return {
-        "id": row["id"],
-        "ts": row["ts"].isoformat() if row["ts"] else None,
-        "user_sub": row["user_sub"],
-        "scope_kind": row["scope_kind"],
-        "scope_id": row["scope_id"],
-        "action": row["action"],
-        "key": row["key"],
-        "target_format": row["target_format"],
-        "status": row["status"],
-        "error": row["error"],
-        "traceback": row["traceback"],
-    }
-
-
-async def mark_audit_log_issue_bot(
-    pool: asyncpg.Pool,
-    audit_id: int,
-    *,
-    status: str,
-    error: str | None = None,
-) -> None:
-    """Stamp a terminal issue-bot status on one audit_log row. Mirrors
-    :func:`mark_audit_run_issue_bot` for the per-row case."""
-    await pool.execute(
-        """
-        UPDATE audit_log
-        SET issue_bot_status = $2,
-            issue_bot_last_error = $3,
-            issue_bot_synced_at = NOW()
-        WHERE id = $1
-        """,
-        audit_id,
-        status,
-        error,
-    )
-
-
-async def reset_audit_log_issue_bot(
-    pool: asyncpg.Pool,
-    audit_id: int,
-) -> bool:
-    """Clear the bot status on one audit_log row so the next tick
-    re-syncs it. Used by the admin "retry sync" button. Returns
-    True on a real reset, False when the row wasn't found or
-    isn't in a terminal failure state (queued/running rows don't
-    have anything to sync)."""
-    result = await pool.execute(
-        """
-        UPDATE audit_log
-        SET issue_bot_status = NULL,
-            issue_bot_last_error = NULL,
-            issue_bot_synced_at = NULL
-        WHERE id = $1 AND status IN ('error', 'failed')
-        """,
-        audit_id,
-    )
-    return result.endswith(" 1")
-
-
-async def aggregate_conversion_metrics(
-    pool: asyncpg.Pool,
-    *,
-    since_days: int = 30,
-    trigger: str | None = None,
-    audit_run_id: str | None = None,
-    worker_image_tag: str | None = None,
-) -> list[dict]:
-    """Per-cell (``source_ext`` × ``target_format``) aggregation over
-    the recent convert jobs (M6 cross-conversion dashboard).
-
-    Computes p50 / p95 / max for duration, peak RSS, RSS per source MB,
-    and write bytes. Failure rate is ``fail_count / sample_count`` —
-    a float in ``[0, 1]``. ``source_size_mb`` is derived from
-    ``read_bytes`` (the storage bytes the worker pulled in); rows with
-    a NULL read_bytes contribute to sample/duration metrics but not
-    to RSS-per-MB.
-
-    ``trigger`` filters the underlying rows:
-      * ``None`` / ``'all'`` — every convert job (default)
-      * ``'audit'`` — only jobs tied to an audit run (M1+ sweeps)
-      * ``'user'`` — only direct user-driven convert jobs
-
-    ``audit_run_id`` narrows to a single sweep; combined with a
-    ``worker_image_tag`` filter that's the way to lock the dashboard
-    to one set of measurements taken with one worker build so an
-    old/cached row from a different image doesn't dilute the
-    numbers. ``since_days`` is clamped to ``[1, 365]`` so a typo'd
-    multi-year range can't accidentally pin the DB; the admin UI
-    exposes a fixed picker (24h / 7d / 30d / 90d).
-    """
-    days = max(1, min(365, since_days))
-    where_extra = ""
-    args: list = []
-    if trigger == "audit":
-        where_extra += " AND audit_run_id IS NOT NULL"
-    elif trigger == "user":
-        where_extra += " AND audit_run_id IS NULL"
-    # Otherwise no trigger filter ("all").
-    args.append(days)
-    if audit_run_id is not None:
-        args.append(audit_run_id)
-        where_extra += f" AND audit_run_id = ${len(args)}"
-    if worker_image_tag is not None:
-        args.append(worker_image_tag)
-        where_extra += f" AND worker_image_tag = ${len(args)}"
-    sql = f"""
-        WITH convert_jobs AS (
-            SELECT
-                LOWER(SUBSTRING(key FROM '\\.([^.]+)$')) AS source_ext,
-                target_format,
-                status,
-                duration_ms,
-                peak_rss_kb,
-                read_bytes,
-                write_bytes,
-                cpu_user_ms,
-                cpu_sys_ms,
-                -- Effective source size in MB. Floor at 0.001 so
-                -- division by zero can't happen for tiny / unknown
-                -- inputs; the resulting RSS/MB inflation only kicks
-                -- in for files <1 KB which are useless data points
-                -- anyway.
-                GREATEST(COALESCE(read_bytes, 0) / 1048576.0, 0.001) AS source_mb
-            FROM audit_log
-            WHERE action = 'convert'
-              -- ``$1`` is an int (clamped days); multiply against
-              -- the interval literal so we don't need to round-
-              -- trip through a string concat (which asyncpg
-              -- rejects with ``expected str, got int`` because
-              -- ``||`` is the SQL string-concat operator).
-              AND ts > NOW() - ($1 * INTERVAL '1 day')
-              AND target_format IS NOT NULL
-              AND key IS NOT NULL
-              {where_extra}
-        )
-        SELECT
-            source_ext,
-            target_format,
-            COUNT(*) AS sample_count,
-            COUNT(*) FILTER (WHERE status IN ('error', 'failed')) AS fail_count,
-            COUNT(*) FILTER (WHERE status IN ('ok', 'done')) AS ok_count,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY duration_ms) AS duration_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY duration_ms) AS duration_ms_p95,
-            MAX(duration_ms) AS duration_ms_max,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY peak_rss_kb) AS peak_rss_kb_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY peak_rss_kb) AS peak_rss_kb_p95,
-            MAX(peak_rss_kb) AS peak_rss_max_kb,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY peak_rss_kb / source_mb)
-                FILTER (WHERE peak_rss_kb IS NOT NULL AND read_bytes > 0)
-                AS peak_rss_per_source_mb_p95,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY write_bytes) AS write_bytes_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY write_bytes) AS write_bytes_p95,
-            AVG(read_bytes)::bigint AS read_bytes_avg,
-            -- IO/CPU split: how much of wall-clock is spent in CPU
-            -- vs blocked on IO. Computed as SUM(cpu_user_ms +
-            -- cpu_sys_ms) / SUM(duration_ms). Values close to 1
-            -- mean the converter is CPU-bound; values < ~0.3 mean
-            -- most of the wall-clock is spent waiting (S3 reads,
-            -- presigned-URL handshakes, OCC tessellation IO, etc.)
-            -- — those are the "consider streaming or async IO"
-            -- candidates. NULL when no rows had timing.
-            CASE
-                WHEN SUM(duration_ms) > 0
-                  THEN SUM(COALESCE(cpu_user_ms, 0) + COALESCE(cpu_sys_ms, 0))::float
-                       / SUM(duration_ms)::float
-                ELSE NULL
-            END AS cpu_fraction
-        FROM convert_jobs
-        WHERE source_ext IS NOT NULL
-          AND source_ext != ''
-          AND target_format != ''
-        GROUP BY source_ext, target_format
-        ORDER BY source_ext, target_format
-    """
-    rows = await pool.fetch(sql, *args)
-
-    def _f(v) -> float | None:
-        # PERCENTILE_CONT returns NUMERIC which asyncpg gives as
-        # Decimal; convert to float for the JSON layer. None stays
-        # None so the frontend can detect "no data" cleanly.
-        if v is None:
-            return None
-        return float(v)
-
-    def _i(v) -> int | None:
-        if v is None:
-            return None
-        return int(v)
-
-    cells: list[dict] = []
-    for r in rows:
-        sample_count = r["sample_count"] or 0
-        fail_count = r["fail_count"] or 0
-        cells.append(
-            {
-                "source_ext": r["source_ext"] or "",
-                "target_format": r["target_format"] or "",
-                "sample_count": sample_count,
-                "fail_count": fail_count,
-                "ok_count": r["ok_count"] or 0,
-                "failure_rate": (fail_count / sample_count if sample_count > 0 else 0.0),
-                "duration_ms_p50": _i(r["duration_ms_p50"]),
-                "duration_ms_p95": _i(r["duration_ms_p95"]),
-                "duration_ms_max": _i(r["duration_ms_max"]),
-                "peak_rss_kb_p50": _i(r["peak_rss_kb_p50"]),
-                "peak_rss_kb_p95": _i(r["peak_rss_kb_p95"]),
-                "peak_rss_max_kb": _i(r["peak_rss_max_kb"]),
-                "peak_rss_per_source_mb_p95": _f(r["peak_rss_per_source_mb_p95"]),
-                "write_bytes_p50": _i(r["write_bytes_p50"]),
-                "write_bytes_p95": _i(r["write_bytes_p95"]),
-                "read_bytes_avg": _i(r["read_bytes_avg"]),
-                "cpu_fraction": _f(r["cpu_fraction"]),
-            }
-        )
-    return cells
-
-
-async def aggregate_view_load_metrics(
-    pool: asyncpg.Pool,
-    *,
-    since_days: int = 30,
-) -> list[dict]:
-    """Per-file aggregation over recent browser model loads (``action =
-    'view'`` rows written by the viewer's opt-in load instrumentation).
-
-    Grouped by ``key`` (the GLB object loaded) because the operational
-    question is "which models are slow to load, and why". For each file
-    it computes p50/p95 of every load phase so a slow load can be
-    attributed to a bottleneck class:
-
-      * IO / backend storage — ``ttfb_ms`` (request -> first byte)
-      * network transfer     — ``download_ms``, throughput, wire bytes
-      * client CPU           — ``decompress_ms`` + ``parse_ms`` + ``prepare_ms``
-      * GPU upload           — ``first_render_ms``
-
-    ``dominant_bound`` labels the row by whichever class holds the
-    largest share of the median total — the load-side analogue of the
-    conversion dashboard's ``cpu_fraction`` signal. Phase fields are
-    pulled out of the ``client_metrics`` JSONB; rows missing a field
-    (e.g. cross-origin presigned loads with no Resource Timing split)
-    simply don't contribute to that field's percentile.
-    """
-    days = max(1, min(365, since_days))
-
-    def num(field: str) -> str:
-        # NULLIF guards an empty-string JSON value from a bad ::numeric cast.
-        return f"NULLIF(client_metrics->>'{field}', '')::numeric"
-
-    sql = f"""
-        WITH loads AS (
-            SELECT
-                key,
-                LOWER(SUBSTRING(key FROM '\\.([^.]+)$')) AS source_ext,
-                status,
-                duration_ms,
-                read_bytes,
-                write_bytes,
-                peak_rss_kb,
-                {num('ttfb_ms')}         AS ttfb_ms,
-                {num('download_ms')}     AS download_ms,
-                {num('decompress_ms')}   AS decompress_ms,
-                {num('parse_ms')}        AS parse_ms,
-                {num('prepare_ms')}      AS prepare_ms,
-                {num('first_render_ms')} AS first_render_ms,
-                {num('total_ms')}        AS total_ms,
-                {num('throughput_mbps')} AS throughput_mbps,
-                {num('transfer_bytes')}  AS transfer_bytes,
-                {num('triangles')}       AS triangles,
-                {num('vertices')}        AS vertices
-            FROM audit_log
-            WHERE action = 'view'
-              AND client_metrics IS NOT NULL
-              AND key IS NOT NULL
-              AND ts > NOW() - ($1 * INTERVAL '1 day')
-        )
-        SELECT
-            key,
-            MAX(source_ext) AS source_ext,
-            COUNT(*) AS sample_count,
-            COUNT(*) FILTER (WHERE status IN ('error', 'failed')) AS fail_count,
-            COUNT(*) FILTER (WHERE status IN ('ok', 'done')) AS ok_count,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY duration_ms) AS total_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY duration_ms) AS total_ms_p95,
-            MAX(duration_ms) AS total_ms_max,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY ttfb_ms) AS ttfb_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY ttfb_ms) AS ttfb_ms_p95,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY download_ms) AS download_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY download_ms) AS download_ms_p95,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY decompress_ms) AS decompress_ms_p50,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY parse_ms) AS parse_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY parse_ms) AS parse_ms_p95,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY prepare_ms) AS prepare_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY prepare_ms) AS prepare_ms_p95,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY first_render_ms) AS first_render_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY first_render_ms) AS first_render_ms_p95,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY throughput_mbps) AS throughput_mbps_p50,
-            AVG(transfer_bytes)::bigint AS transfer_bytes_avg,
-            AVG(read_bytes)::bigint AS read_bytes_avg,
-            AVG(write_bytes)::bigint AS write_bytes_avg,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY triangles) AS triangles_p50,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY vertices) AS vertices_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY peak_rss_kb) AS js_heap_kb_p95
-        FROM loads
-        GROUP BY key
-        ORDER BY total_ms_p95 DESC NULLS LAST, key
-    """
-    rows = await pool.fetch(sql, days)
-
-    def _f(v) -> float | None:
-        return float(v) if v is not None else None
-
-    def _i(v) -> int | None:
-        return int(v) if v is not None else None
-
-    cells: list[dict] = []
-    for r in rows:
-        sample_count = r["sample_count"] or 0
-        fail_count = r["fail_count"] or 0
-        # Bottleneck attribution from the median phases. CPU is the sum
-        # of the three main-thread phases; whichever class holds the
-        # biggest slice of the median total wins. Fall back to "unknown"
-        # when there's no timing at all.
-        io = _f(r["ttfb_ms_p50"]) or 0.0
-        net = _f(r["download_ms_p50"]) or 0.0
-        cpu = (_f(r["decompress_ms_p50"]) or 0.0) + (_f(r["parse_ms_p50"]) or 0.0) + (_f(r["prepare_ms_p50"]) or 0.0)
-        gpu = _f(r["first_render_ms_p50"]) or 0.0
-        classes = {"io": io, "network": net, "cpu": cpu, "gpu": gpu}
-        accounted = sum(classes.values())
-        dominant = max(classes, key=classes.get) if accounted > 0 else "unknown"
-        cells.append(
-            {
-                "key": r["key"] or "",
-                "source_ext": r["source_ext"] or "",
-                "sample_count": sample_count,
-                "fail_count": fail_count,
-                "ok_count": r["ok_count"] or 0,
-                "failure_rate": (fail_count / sample_count if sample_count > 0 else 0.0),
-                "total_ms_p50": _i(r["total_ms_p50"]),
-                "total_ms_p95": _i(r["total_ms_p95"]),
-                "total_ms_max": _i(r["total_ms_max"]),
-                "ttfb_ms_p50": _i(r["ttfb_ms_p50"]),
-                "ttfb_ms_p95": _i(r["ttfb_ms_p95"]),
-                "download_ms_p50": _i(r["download_ms_p50"]),
-                "download_ms_p95": _i(r["download_ms_p95"]),
-                "decompress_ms_p50": _i(r["decompress_ms_p50"]),
-                "parse_ms_p50": _i(r["parse_ms_p50"]),
-                "parse_ms_p95": _i(r["parse_ms_p95"]),
-                "prepare_ms_p50": _i(r["prepare_ms_p50"]),
-                "prepare_ms_p95": _i(r["prepare_ms_p95"]),
-                "first_render_ms_p50": _i(r["first_render_ms_p50"]),
-                "first_render_ms_p95": _i(r["first_render_ms_p95"]),
-                "throughput_mbps_p50": _f(r["throughput_mbps_p50"]),
-                "transfer_bytes_avg": _i(r["transfer_bytes_avg"]),
-                "read_bytes_avg": _i(r["read_bytes_avg"]),
-                "write_bytes_avg": _i(r["write_bytes_avg"]),
-                "triangles_p50": _i(r["triangles_p50"]),
-                "vertices_p50": _i(r["vertices_p50"]),
-                "js_heap_kb_p95": _i(r["js_heap_kb_p95"]),
-                # Median-phase bottleneck attribution.
-                "io_ms": round(io, 1),
-                "network_ms": round(net, 1),
-                "cpu_ms": round(cpu, 1),
-                "gpu_ms": round(gpu, 1),
-                "dominant_bound": dominant,
-            }
-        )
-    return cells
-
-
-async def aggregate_view_load_hotspots(
-    pool: asyncpg.Pool,
-    *,
-    action: str = "view",
-    key: str | None = None,
-    since_days: int = 30,
-    limit: int = 100,
-) -> dict:
-    """Function-level hotspots across browser ``view`` loads or ``render``
-    windows — the client-side analogue of ``aggregate_profile_hotspots``
-    for conversions.
-
-    The viewer's opt-in instrumentation runs the JS Self-Profiling API
-    (during a load, or per render window when ``action='render'``) and
-    stores the top self-time frames (TypeScript *and* WASM — pyodide/
-    adacpp frames surface as ``wasm-function[...]`` or their name-section
-    names) under ``client_metrics->'profile_frames'`` as
-    ``[{"fn", "self_ms", "total_ms"}, ...]``. This unnests them across
-    every matching row (optionally one ``key``) and sums self-time per
-    function so the slowest TS/WASM calls float to the top.
-
-    Returns ``{functions: [...], loads_in_window: N}``; an empty
-    ``functions`` with ``loads_in_window=0`` means no profiled rows in
-    the window (self-profiling unsupported/disabled, or the
-    ``Document-Policy: js-profiling`` header isn't being served).
-    """
-    days = max(1, min(365, since_days))
-    lim = max(1, min(1000, limit))
-    act = action if action in ("view", "render") else "view"
-    args: list = [act, days]
-    key_filter = ""
-    if key:
-        args.append(key)
-        key_filter = f" AND key = ${len(args)}"
-
-    # Count profiled rows in the window first (so the UI can distinguish
-    # "no data" from "no hotspots").
-    loads_in_window = await pool.fetchval(
-        f"""
-        SELECT COUNT(*) FROM audit_log
-        WHERE action = $1
-          AND client_metrics ? 'profile_frames'
-          AND jsonb_array_length(client_metrics->'profile_frames') > 0
-          AND ts > NOW() - ($2 * INTERVAL '1 day')
-          {key_filter}
-        """,
-        *args,
-    )
-
-    args.append(lim)
-    rows = await pool.fetch(
-        f"""
-        WITH frames AS (
-            SELECT f->>'fn' AS fn,
-                   NULLIF(f->>'self_ms', '')::numeric AS self_ms,
-                   NULLIF(f->>'total_ms', '')::numeric AS total_ms
-            FROM audit_log,
-                 LATERAL jsonb_array_elements(client_metrics->'profile_frames') AS f
-            WHERE action = $1
-              AND client_metrics ? 'profile_frames'
-              AND ts > NOW() - ($2 * INTERVAL '1 day')
-              {key_filter}
-        )
-        SELECT fn,
-               COUNT(*) AS samples,
-               SUM(self_ms)::numeric AS self_ms_sum,
-               AVG(self_ms)::numeric AS self_ms_avg,
-               MAX(total_ms)::numeric AS total_ms_max
-        FROM frames
-        WHERE fn IS NOT NULL AND fn != ''
-        GROUP BY fn
-        ORDER BY self_ms_sum DESC NULLS LAST
-        LIMIT ${len(args)}
-        """,
-        *args,
-    )
-
-    def _f(v) -> float | None:
-        return float(v) if v is not None else None
-
-    functions = [
-        {
-            "fn": r["fn"],
-            "samples": r["samples"] or 0,
-            "self_ms_sum": _f(r["self_ms_sum"]),
-            "self_ms_avg": _f(r["self_ms_avg"]),
-            "total_ms_max": _f(r["total_ms_max"]),
-            "is_wasm": bool(r["fn"]) and ("wasm" in r["fn"].lower()),
-        }
-        for r in rows
-    ]
-    return {"functions": functions, "loads_in_window": int(loads_in_window or 0)}
-
-
-async def aggregate_render_metrics(
-    pool: asyncpg.Pool,
-    *,
-    since_days: int = 30,
-) -> list[dict]:
-    """Per-file aggregation over steady-state render windows (``action =
-    'render'`` rows). Each row is one rolling window the viewer sampled
-    while a model was on screen; this rolls them up per ``key`` so a
-    janky / GPU-bound model is obvious.
-
-    For each file: median/worst FPS, CPU frame time (time between frames),
-    GPU frame time (``EXT_disjoint_timer_query_webgl2`` when the client
-    had it), draw calls + triangles rendered, and a ``dominant_bound``
-    label — ``gpu`` when median GPU ms exceeds median CPU frame ms, else
-    ``cpu`` — the steady-state analogue of the load dashboard's bound.
-    """
-    days = max(1, min(365, since_days))
-
-    def num(field: str) -> str:
-        return f"NULLIF(client_metrics->>'{field}', '')::numeric"
-
-    sql = f"""
-        WITH r AS (
-            SELECT
-                key,
-                {num('fps_p50')}        AS fps_p50,
-                {num('fps_min')}        AS fps_min,
-                {num('frame_ms_p50')}   AS frame_ms_p50,
-                {num('frame_ms_p95')}   AS frame_ms_p95,
-                {num('gpu_ms_p50')}     AS gpu_ms_p50,
-                {num('gpu_ms_p95')}     AS gpu_ms_p95,
-                {num('draw_calls')}     AS draw_calls,
-                {num('triangles')}      AS triangles,
-                {num('programs')}       AS programs,
-                {num('geometries')}     AS geometries,
-                {num('textures')}       AS textures,
-                {num('long_frames')}    AS long_frames,
-                {num('frame_count')}    AS frame_count
-            FROM audit_log
-            WHERE action = 'render'
-              AND client_metrics IS NOT NULL
-              AND key IS NOT NULL
-              AND ts > NOW() - ($1 * INTERVAL '1 day')
-        )
-        SELECT
-            key,
-            COUNT(*) AS window_count,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY fps_p50) AS fps_p50,
-            MIN(fps_min) AS fps_min,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY frame_ms_p50) AS frame_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY frame_ms_p95) AS frame_ms_p95,
-            PERCENTILE_CONT(0.5)  WITHIN GROUP (ORDER BY gpu_ms_p50) AS gpu_ms_p50,
-            PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY gpu_ms_p95) AS gpu_ms_p95,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY draw_calls) AS draw_calls_p50,
-            MAX(draw_calls) AS draw_calls_max,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY triangles) AS triangles_p50,
-            MAX(programs) AS programs_max,
-            MAX(geometries) AS geometries_max,
-            MAX(textures) AS textures_max,
-            SUM(long_frames) AS long_frames_sum,
-            SUM(frame_count) AS frame_count_sum
-        FROM r
-        GROUP BY key
-        ORDER BY fps_p50 ASC NULLS LAST, key
-    """
-    rows = await pool.fetch(sql, days)
-
-    def _f(v) -> float | None:
-        return float(v) if v is not None else None
-
-    def _i(v) -> int | None:
-        return int(v) if v is not None else None
-
-    cells: list[dict] = []
-    for r in rows:
-        cpu_ms = _f(r["frame_ms_p50"])
-        gpu_ms = _f(r["gpu_ms_p50"])
-        if gpu_ms is not None and cpu_ms is not None:
-            dominant = "gpu" if gpu_ms > cpu_ms else "cpu"
-        elif gpu_ms is not None:
-            dominant = "gpu"
-        elif cpu_ms is not None:
-            dominant = "cpu"
-        else:
-            dominant = "unknown"
-        cells.append(
-            {
-                "key": r["key"] or "",
-                "window_count": r["window_count"] or 0,
-                "fps_p50": _f(r["fps_p50"]),
-                "fps_min": _f(r["fps_min"]),
-                "frame_ms_p50": _f(r["frame_ms_p50"]),
-                "frame_ms_p95": _f(r["frame_ms_p95"]),
-                "gpu_ms_p50": _f(r["gpu_ms_p50"]),
-                "gpu_ms_p95": _f(r["gpu_ms_p95"]),
-                "draw_calls_p50": _i(r["draw_calls_p50"]),
-                "draw_calls_max": _i(r["draw_calls_max"]),
-                "triangles_p50": _i(r["triangles_p50"]),
-                "programs_max": _i(r["programs_max"]),
-                "geometries_max": _i(r["geometries_max"]),
-                "textures_max": _i(r["textures_max"]),
-                "long_frames_sum": _i(r["long_frames_sum"]),
-                "frame_count_sum": _i(r["frame_count_sum"]),
-                "dominant_bound": dominant,
-            }
-        )
-    return cells
-
-
-async def get_audit_client_metrics(pool: asyncpg.Pool, audit_id: int) -> dict | None:
-    """Return the ``client_metrics`` JSONB for one audit_log row (the
-    browser view/render instrumentation payload), or None. Backs the
-    admin audit-log detail view so a single load/render event can be
-    inspected phase-by-phase."""
-    row = await pool.fetchrow("SELECT client_metrics FROM audit_log WHERE id = $1", audit_id)
-    if row is None:
-        return None
-    cm = row["client_metrics"]
-    if cm is None:
-        return None
-    # JSONB comes back as text (no codec set); parse defensively, mirroring
-    # the metrics_samples / counts read paths.
-    if isinstance(cm, str):
-        try:
-            cm = json.loads(cm)
-        except Exception:
-            return None
-    return cm if isinstance(cm, dict) else None
-
-
-# ── Profile hotspots (M7 perf dashboard) ────────────────────────────
-
-
-async def claim_unprocessed_profile_row(
-    pool: asyncpg.Pool,
-) -> dict | None:
-    """Atomically claim the oldest audit_log row whose ``.prof`` has
-    not been processed yet by the profile-hotspots background loop.
-
-    Stamps ``profile_stats_processed_at`` so a concurrent replica
-    skips the row; the caller is expected to either insert its
-    parsed function stats or (on failure) overwrite the same
-    timestamp via :func:`mark_profile_stats_processed` with an
-    error message. Returns the audit_log row's
-    ``id`` / ``key`` / ``target_format`` / ``profile_key`` /
-    ``scope_kind`` / ``scope_id`` so the parser can fetch the blob
-    without a second lookup.
-
-    Limits to ``status IN ('ok', 'done')`` — only completed
-    conversions have a meaningful profile; queued / failed cells
-    either never produced one or produced a partial dump we don't
-    want polluting the aggregates.
-    """
-    row = await pool.fetchrow(
-        """
-        UPDATE audit_log
-        SET profile_stats_processed_at = NOW()
-        WHERE id = (
-            SELECT id FROM audit_log
-            WHERE profile_key IS NOT NULL
-              AND profile_stats_processed_at IS NULL
-              AND status IN ('ok', 'done')
-            ORDER BY id ASC
-            LIMIT 1
-            FOR UPDATE SKIP LOCKED
-        )
-        RETURNING id, key, target_format, profile_key, scope_kind, scope_id
-        """,
-    )
-    if row is None:
-        return None
-    return {
-        "id": row["id"],
-        "key": row["key"],
-        "target_format": row["target_format"],
-        "profile_key": row["profile_key"],
-        "scope_kind": row["scope_kind"],
-        "scope_id": row["scope_id"],
-    }
-
-
-async def insert_profile_function_stats(
-    pool: asyncpg.Pool,
-    audit_id: int,
-    rows: list[dict],
-) -> None:
-    """Insert the parsed top-K function stats for one audit row.
-
-    ``rows`` is the output of pstats parsing already truncated to
-    K=50 (or whatever the caller picked) and sorted by ``cumtime``
-    desc. ``rank`` is the row's index in that order. Uses a single
-    executemany so the K inserts don't open K transactions.
-    """
-    if not rows:
-        return
-    values = [
-        (
-            audit_id,
-            idx,
-            r["func"],
-            r["file"],
-            r["line"],
-            int(r["ncalls"]),
-            int(r["primitive_calls"]),
-            float(r["tottime"]),
-            float(r["cumtime"]),
-        )
-        for idx, r in enumerate(rows)
-    ]
-    await pool.executemany(
-        """
-        INSERT INTO profile_function_stats
-            (audit_id, rank, func, file, line, ncalls,
-             primitive_calls, tottime, cumtime)
-        VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
-        """,
-        values,
-    )
-
-
-async def mark_profile_stats_failed(
-    pool: asyncpg.Pool,
-    audit_id: int,
-    error: str,
-) -> None:
-    """Stamp a parse failure on the audit_log row. The timestamp is
-    already set by :func:`claim_unprocessed_profile_row` (the
-    claim doubles as a "we touched this row" marker so it doesn't
-    re-claim on the next tick); we just attach the error message
-    so the admin UI can show what went wrong."""
-    await pool.execute(
-        "UPDATE audit_log SET profile_stats_error = $2 WHERE id = $1",
-        audit_id,
-        error,
-    )
-
-
-async def aggregate_profile_hotspots(
-    pool: asyncpg.Pool,
-    *,
-    source_ext: str | None = None,
-    target_format: str | None = None,
-    since_days: int = 30,
-    limit: int = 25,
-) -> dict:
-    """Cross-run hotspot aggregation for one (source_ext, target_format)
-    cell.
-
-    Joins ``profile_function_stats`` against ``audit_log`` to filter
-    by source-extension + target + time window, GROUPs by
-    ``(func, file, line)`` and SUMs cumulative time + call count
-    across every profile that landed in the window.
-
-    Returns:
-        {"functions": [{"func", "file", "line", "agg_cumtime",
-                        "agg_ncalls", "profiles_seen"}],
-         "profiles_in_window": N,
-         "total_cumtime_in_window": T,
-         "since_days": N}
-    """
-    days = max(1, min(365, since_days))
-    where = [
-        "al.action = 'convert'",
-        "al.ts > NOW() - ($1 * INTERVAL '1 day')",
-    ]
-    args: list = [days]
-    if source_ext:
-        ext = source_ext.lower()
-        # Strip leading dot if present; we match against the SUBSTRING
-        # pattern that already excludes it.
-        ext_no_dot = ext.lstrip(".")
-        args.append(ext_no_dot)
-        where.append(f"LOWER(SUBSTRING(al.key FROM '\\.([^.]+)$')) = ${len(args)}")
-    if target_format:
-        args.append(target_format.lower())
-        where.append(f"LOWER(al.target_format) = ${len(args)}")
-    args.append(max(1, min(500, limit)))
-    sql = f"""
-        WITH cell_rows AS (
-            SELECT al.id, al.duration_ms
-            FROM audit_log al
-            WHERE {' AND '.join(where)}
-              AND al.profile_key IS NOT NULL
-              AND al.profile_stats_processed_at IS NOT NULL
-        ),
-        agg AS (
-            SELECT
-                pfs.func, pfs.file, pfs.line,
-                SUM(pfs.cumtime) AS agg_cumtime,
-                SUM(pfs.ncalls) AS agg_ncalls,
-                COUNT(DISTINCT pfs.audit_id) AS profiles_seen
-            FROM profile_function_stats pfs
-            JOIN cell_rows c ON c.id = pfs.audit_id
-            GROUP BY pfs.func, pfs.file, pfs.line
-        )
-        SELECT * FROM agg
-        ORDER BY agg_cumtime DESC
-        LIMIT ${len(args)}
-    """
-    rows = await pool.fetch(sql, *args)
-
-    # Count profiles + window total cumtime separately so the UI
-    # can show "top N of M profiles" without scanning the join.
-    counts = await pool.fetchrow(
-        f"""
-        SELECT
-            COUNT(DISTINCT pfs.audit_id) AS profiles_in_window,
-            COALESCE(SUM(pfs.cumtime) FILTER (WHERE pfs.rank = 0), 0)
-                AS total_top_cumtime_in_window
-        FROM profile_function_stats pfs
-        JOIN audit_log al ON al.id = pfs.audit_id
-        WHERE {' AND '.join(where)}
-          AND al.profile_stats_processed_at IS NOT NULL
-        """,
-        *args[:-1],  # drop the limit param — counts query doesn't use it
-    )
-
-    return {
-        "functions": [
-            {
-                "func": r["func"],
-                "file": r["file"],
-                "line": r["line"],
-                "agg_cumtime": float(r["agg_cumtime"]),
-                "agg_ncalls": int(r["agg_ncalls"]),
-                "profiles_seen": int(r["profiles_seen"]),
-            }
-            for r in rows
-        ],
-        "profiles_in_window": int(counts["profiles_in_window"] or 0),
-        "total_top_cumtime_in_window": float(counts["total_top_cumtime_in_window"] or 0.0),
-        "since_days": days,
-    }
-
-
-async def list_failed_audit_run_jobs(
-    pool: asyncpg.Pool,
-    run_id: str,
-) -> list[dict]:
-    """All ``audit_log`` rows in ``run_id`` whose status indicates a
-    failure ('error' or 'failed'). Returns the columns the issue-bot
-    needs to fingerprint + describe the failure — key, target,
-    error message, traceback excerpt. Cached cells (status='done')
-    and queued cells that never resolved aren't included; the bot
-    only opens issues for real failures."""
-    rows = await pool.fetch(
-        """
-        SELECT id, key, scope_kind, scope_id, target_format,
-               status, error, traceback
-        FROM audit_log
-        WHERE audit_run_id = $1
-          AND status IN ('error', 'failed')
-        ORDER BY id ASC
-        """,
-        run_id,
-    )
-    return [
-        {
-            "id": r["id"],
-            "key": r["key"],
-            "scope_kind": r["scope_kind"],
-            "scope_id": r["scope_id"],
-            "target_format": r["target_format"],
-            "status": r["status"],
-            "error": r["error"],
-            "traceback": r["traceback"],
-        }
-        for r in rows
-    ]
-
-
-# ── Procedural cell models ───────────────────────────────────────────
-
-
-def _procedural_row_summary(r) -> dict:
-    out = {
-        "id": str(r["id"]),
-        "name": r["name"],
-        "revision": r["revision"],
-        "created_by": r["created_by"],
-        "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-        "updated_at": r["updated_at"].isoformat() if r["updated_at"] else None,
-    }
-    # The engine/schema_version columns were added later (migration 026); guard so
-    # a summary built from a SELECT that omits them doesn't KeyError.
-    if "engine" in r:
-        out["engine"] = r["engine"]
-    if "schema_version" in r:
-        out["schema_version"] = r["schema_version"]
-    return out
-
-
-async def create_procedural_model(
-    pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None, name: str, created_by: str | None
-) -> dict | None:
-    """Insert a new (empty) procedural model. Returns the full row incl. doc,
-    or None when a live model with that name already exists in the scope."""
-    try:
-        row = await pool.fetchrow(
-            """
-            INSERT INTO procedural_models (scope_kind, scope_id, name, created_by)
-            VALUES ($1, $2, $3, $4)
-            RETURNING id, name, doc, revision, engine, schema_version, created_by, created_at, updated_at
-            """,
-            scope_kind,
-            scope_id,
-            name,
-            created_by,
-        )
-    except asyncpg.UniqueViolationError:
-        return None
-    out = _procedural_row_summary(row)
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def rename_procedural_model(pool: asyncpg.Pool, model_id: str, name: str) -> dict | None | bool:
-    """Rename a model, which is also how it MOVES between folders.
-
-    The name carries the folder path (see procedural.normalize_model_name), so
-    there is one operation rather than two that could disagree about where a
-    model lives.
-
-    Returns the updated summary, ``False`` when the target name is taken (the
-    scope-unique index — the same collision a filesystem would report), or
-    ``None`` when the model does not exist or is archived.
-    """
-    try:
-        row = await pool.fetchrow(
-            """
-            UPDATE procedural_models
-            SET name = $2, updated_at = now()
-            WHERE id = $1::uuid AND NOT archived
-            RETURNING id, name, doc, revision, engine, schema_version, created_by, created_at, updated_at
-            """,
-            model_id,
-            name,
-        )
-    except asyncpg.UniqueViolationError:
-        return False
-    if row is None:
-        return None
-    return _procedural_row_summary(row)
-
-
-async def list_procedural_models(pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None) -> list[dict]:
-    rows = await pool.fetch(
-        """
-        SELECT id, name, revision, engine, schema_version, created_by, created_at, updated_at
-        FROM procedural_models
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '') AND NOT archived
-        ORDER BY name ASC
-        """,
-        scope_kind,
-        scope_id,
-    )
-    return [_procedural_row_summary(r) for r in rows]
-
-
-async def get_procedural_model(pool: asyncpg.Pool, model_id: str) -> dict | None:
-    row = await pool.fetchrow(
-        """
-        SELECT id, scope_kind, scope_id, name, doc, revision, engine, schema_version,
-               created_by, created_at, updated_at
-        FROM procedural_models
-        WHERE id = $1 AND NOT archived
-        """,
-        model_id,
-    )
-    if row is None:
-        return None
-    out = _procedural_row_summary(row)
-    out["scope_kind"] = row["scope_kind"]
-    out["scope_id"] = row["scope_id"]
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def update_procedural_model_doc(pool: asyncpg.Pool, model_id: str, doc: dict, base_revision: int) -> int | None:
-    """Optimistic-concurrency doc update: bumps revision only when the caller's
-    base_revision matches. Returns the new revision, or None on conflict. The
-    engine/schema_version columns are mirrored from the doc's routing header so
-    they stay the single source of truth."""
-    row = await pool.fetchrow(
-        """
-        UPDATE procedural_models
-        SET doc = $2::jsonb, revision = revision + 1, updated_at = now(),
-            engine = COALESCE($4, engine), schema_version = COALESCE($5, schema_version)
-        WHERE id = $1 AND revision = $3 AND NOT archived
-        RETURNING revision
-        """,
-        model_id,
-        json.dumps(doc),
-        base_revision,
-        doc.get("engine"),
-        doc.get("schema_version"),
-    )
-    return None if row is None else row["revision"]
-
-
-async def archive_procedural_model(pool: asyncpg.Pool, model_id: str) -> bool:
-    res = await pool.execute(
-        "UPDATE procedural_models SET archived = TRUE, updated_at = now() WHERE id = $1 AND NOT archived",
-        model_id,
-    )
-    return res.endswith("1")
-
-
-# ── Equipment-type catalog (per-scope) ───────────────────────────────
-
-
-def _equipment_row_summary(r) -> dict:
-    return {
-        "id": str(r["id"]),
-        "slug": r["slug"],
-        "name": r["name"],
-        "description": r["description"],
-        "cad_key": r["cad_key"],
-        "revision": r["revision"],
-        "created_by": r["created_by"],
-        "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-        "updated_at": r["updated_at"].isoformat() if r["updated_at"] else None,
-    }
-
-
-async def create_equipment_type(
-    pool: asyncpg.Pool,
-    *,
-    scope_kind: str,
-    scope_id: str | None,
-    slug: str,
-    name: str,
-    description: str | None,
-    created_by: str | None,
-) -> dict | None:
-    """Insert a new equipment type (default doc). Returns the full row incl.
-    doc, or None when a live type with that slug already exists in the scope."""
-    try:
-        row = await pool.fetchrow(
-            """
-            INSERT INTO equipment_types (scope_kind, scope_id, slug, name, description, created_by)
-            VALUES ($1, $2, $3, $4, $5, $6)
-            RETURNING id, slug, name, description, doc, cad_key, revision, created_by, created_at, updated_at
-            """,
-            scope_kind,
-            scope_id,
-            slug,
-            name,
-            description,
-            created_by,
-        )
-    except asyncpg.UniqueViolationError:
-        return None
-    out = _equipment_row_summary(row)
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def list_equipment_types(pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None) -> list[dict]:
-    rows = await pool.fetch(
-        """
-        SELECT id, slug, name, description, cad_key, revision, created_by, created_at, updated_at
-        FROM equipment_types
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '') AND NOT archived
-        ORDER BY name ASC
-        """,
-        scope_kind,
-        scope_id,
-    )
-    return [_equipment_row_summary(r) for r in rows]
-
-
-async def get_equipment_docs_by_scope(pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None) -> dict[str, dict]:
-    """Map ``slug -> doc`` for the live equipment types in a scope. Used by the
-    procedural compiler to resolve a placed catalog equipment to its
-    bbox/mass/ports definition."""
-    rows = await pool.fetch(
-        """
-        SELECT slug, doc FROM equipment_types
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '') AND NOT archived
-        """,
-        scope_kind,
-        scope_id,
-    )
-    return {r["slug"]: _loads_jsonb(r["doc"]) for r in rows}
-
-
-async def get_equipment_cad_keys_by_scope(
-    pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None
-) -> dict[str, str]:
-    """Map ``slug -> cad_key`` for live equipment types in a scope that have a
-    linked CAD asset. Used to splice real CAD geometry into a compiled model."""
-    rows = await pool.fetch(
-        """
-        SELECT slug, cad_key FROM equipment_types
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '')
-          AND NOT archived AND cad_key IS NOT NULL
-        """,
-        scope_kind,
-        scope_id,
-    )
-    return {r["slug"]: r["cad_key"] for r in rows}
-
-
-async def get_catalog_fingerprint(pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None) -> str:
-    """A short, stable fingerprint of a scope's EQUIPMENT + SYSTEM catalogs — the
-    live inputs a procedural compile resolves that are NOT captured by the model's
-    own revision. Aggregates every non-archived catalog item's ``slug:revision``
-    (plus ``cad_key`` for equipment, since re-linking CAD changes the spliced
-    geometry). The value therefore changes whenever a catalog item is added (a new
-    slug appears), edited (its revision bumps), CAD-relinked, or archived (its slug
-    drops out). Folded into the procedural compile cache decision (via a sidecar) so
-    a catalog edit forces a fresh compile instead of serving the revision-stamped
-    stale artifact. Empty catalogs hash to a stable constant (md5 of '')."""
-    eq_fp = await pool.fetchval(
-        """
-        SELECT md5(COALESCE(
-            string_agg(slug || ':' || revision::text || ':' || COALESCE(cad_key, ''), ',' ORDER BY slug),
-            ''))
-        FROM equipment_types
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '') AND NOT archived
-        """,
-        scope_kind,
-        scope_id,
-    )
-    sy_fp = await pool.fetchval(
-        """
-        SELECT md5(COALESCE(string_agg(slug || ':' || revision::text, ',' ORDER BY slug), ''))
-        FROM system_templates
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '') AND NOT archived
-        """,
-        scope_kind,
-        scope_id,
-    )
-    return hashlib.sha256(f"eq:{eq_fp}|sy:{sy_fp}".encode("utf-8")).hexdigest()[:16]
-
-
-async def get_equipment_type(pool: asyncpg.Pool, type_id: str) -> dict | None:
-    row = await pool.fetchrow(
-        """
-        SELECT id, scope_kind, scope_id, slug, name, description, doc, cad_key,
-               revision, created_by, created_at, updated_at
-        FROM equipment_types
-        WHERE id = $1 AND NOT archived
-        """,
-        type_id,
-    )
-    if row is None:
-        return None
-    out = _equipment_row_summary(row)
-    out["scope_kind"] = row["scope_kind"]
-    out["scope_id"] = row["scope_id"]
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def update_equipment_type(
-    pool: asyncpg.Pool,
-    type_id: str,
-    *,
-    slug: str,
-    name: str,
-    description: str | None,
-    doc: dict,
-    base_revision: int,
-) -> int | None:
-    """Optimistic-concurrency update of an equipment type's metadata + doc.
-    Returns the new revision, or None on revision conflict. Propagates
-    asyncpg.UniqueViolationError when the new slug collides in-scope."""
-    row = await pool.fetchrow(
-        """
-        UPDATE equipment_types
-        SET slug = $2, name = $3, description = $4, doc = $5::jsonb,
-            revision = revision + 1, updated_at = now()
-        WHERE id = $1 AND revision = $6 AND NOT archived
-        RETURNING revision
-        """,
-        type_id,
-        slug,
-        name,
-        description,
-        json.dumps(doc),
-        base_revision,
-    )
-    return None if row is None else row["revision"]
-
-
-async def set_equipment_type_cad(pool: asyncpg.Pool, type_id: str, cad_key: str | None) -> bool:
-    """Link (or clear) the equipment type's CAD asset. Orthogonal to the doc
-    revision — the preview key is not revision-stamped."""
-    res = await pool.execute(
-        "UPDATE equipment_types SET cad_key = $2, updated_at = now() WHERE id = $1 AND NOT archived",
-        type_id,
-        cad_key,
-    )
-    return res.endswith("1")
-
-
-async def apply_inferred_bbox(pool: asyncpg.Pool, type_id: str, bbox: dict) -> bool:
-    """Merge a worker-inferred bounding box into the doc without bumping the
-    revision (a background inference must not race the user's optimistic
-    commits)."""
-    res = await pool.execute(
-        "UPDATE equipment_types SET doc = jsonb_set(doc, '{bbox}', $2::jsonb, true), "
-        "updated_at = now() WHERE id = $1 AND NOT archived",
-        type_id,
-        json.dumps(bbox),
-    )
-    return res.endswith("1")
-
-
-async def archive_equipment_type(pool: asyncpg.Pool, type_id: str) -> bool:
-    res = await pool.execute(
-        "UPDATE equipment_types SET archived = TRUE, updated_at = now() WHERE id = $1 AND NOT archived",
-        type_id,
-    )
-    return res.endswith("1")
-
-
-# ── System-template catalog (per-scope) ──────────────────────────────
-
-
-def _system_row_summary(r) -> dict:
-    return {
-        "id": str(r["id"]),
-        "slug": r["slug"],
-        "name": r["name"],
-        "description": r["description"],
-        "revision": r["revision"],
-        "created_by": r["created_by"],
-        "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-        "updated_at": r["updated_at"].isoformat() if r["updated_at"] else None,
-    }
-
-
-async def create_system_template(
-    pool: asyncpg.Pool,
-    *,
-    scope_kind: str,
-    scope_id: str | None,
-    slug: str,
-    name: str,
-    description: str | None,
-    created_by: str | None,
-) -> dict | None:
-    try:
-        row = await pool.fetchrow(
-            """
-            INSERT INTO system_templates (scope_kind, scope_id, slug, name, description, created_by)
-            VALUES ($1, $2, $3, $4, $5, $6)
-            RETURNING id, slug, name, description, doc, revision, created_by, created_at, updated_at
-            """,
-            scope_kind,
-            scope_id,
-            slug,
-            name,
-            description,
-            created_by,
-        )
-    except asyncpg.UniqueViolationError:
-        return None
-    out = _system_row_summary(row)
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def list_system_templates(pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None) -> list[dict]:
-    rows = await pool.fetch(
-        """
-        SELECT id, slug, name, description, revision, created_by, created_at, updated_at
-        FROM system_templates
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '') AND NOT archived
-        ORDER BY name ASC
-        """,
-        scope_kind,
-        scope_id,
-    )
-    return [_system_row_summary(r) for r in rows]
-
-
-async def get_system_template(pool: asyncpg.Pool, template_id: str) -> dict | None:
-    row = await pool.fetchrow(
-        """
-        SELECT id, scope_kind, scope_id, slug, name, description, doc,
-               revision, created_by, created_at, updated_at
-        FROM system_templates
-        WHERE id = $1 AND NOT archived
-        """,
-        template_id,
-    )
-    if row is None:
-        return None
-    out = _system_row_summary(row)
-    out["scope_kind"] = row["scope_kind"]
-    out["scope_id"] = row["scope_id"]
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def update_system_template(
-    pool: asyncpg.Pool,
-    template_id: str,
-    *,
-    slug: str,
-    name: str,
-    description: str | None,
-    doc: dict,
-    base_revision: int,
-) -> int | None:
-    row = await pool.fetchrow(
-        """
-        UPDATE system_templates
-        SET slug = $2, name = $3, description = $4, doc = $5::jsonb,
-            revision = revision + 1, updated_at = now()
-        WHERE id = $1 AND revision = $6 AND NOT archived
-        RETURNING revision
-        """,
-        template_id,
-        slug,
-        name,
-        description,
-        json.dumps(doc),
-        base_revision,
-    )
-    return None if row is None else row["revision"]
-
-
-async def archive_system_template(pool: asyncpg.Pool, template_id: str) -> bool:
-    res = await pool.execute(
-        "UPDATE system_templates SET archived = TRUE, updated_at = now() WHERE id = $1 AND NOT archived",
-        template_id,
-    )
-    return res.endswith("1")
-
-
-# ── Procedural-engine registry (per-scope) ───────────────────────────
-
-
-def _engine_row_summary(r) -> dict:
-    return {
-        "id": str(r["id"]),
-        "slug": r["slug"],
-        "name": r["name"],
-        "description": r["description"],
-        "revision": r["revision"],
-        "created_by": r["created_by"],
-        "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-        "updated_at": r["updated_at"].isoformat() if r["updated_at"] else None,
-    }
-
-
-async def create_procedural_engine(
-    pool: asyncpg.Pool,
-    *,
-    scope_kind: str,
-    scope_id: str | None,
-    slug: str,
-    name: str,
-    description: str | None,
-    created_by: str | None,
-) -> dict | None:
-    """Insert a new procedural engine (default builtin doc). Returns the full row
-    incl. doc, or None when a live engine with that slug already exists."""
-    try:
-        row = await pool.fetchrow(
-            """
-            INSERT INTO procedural_engines (scope_kind, scope_id, slug, name, description, created_by)
-            VALUES ($1, $2, $3, $4, $5, $6)
-            RETURNING id, slug, name, description, doc, revision, created_by, created_at, updated_at
-            """,
-            scope_kind,
-            scope_id,
-            slug,
-            name,
-            description,
-            created_by,
-        )
-    except asyncpg.UniqueViolationError:
-        return None
-    out = _engine_row_summary(row)
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def list_procedural_engines(pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None) -> list[dict]:
-    rows = await pool.fetch(
-        """
-        SELECT id, slug, name, description, revision, created_by, created_at, updated_at
-        FROM procedural_engines
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '') AND NOT archived
-        ORDER BY name ASC
-        """,
-        scope_kind,
-        scope_id,
-    )
-    return [_engine_row_summary(r) for r in rows]
-
-
-async def get_procedural_engine(pool: asyncpg.Pool, engine_id: str) -> dict | None:
-    row = await pool.fetchrow(
-        """
-        SELECT id, scope_kind, scope_id, slug, name, description, doc,
-               revision, created_by, created_at, updated_at
-        FROM procedural_engines
-        WHERE id = $1 AND NOT archived
-        """,
-        engine_id,
-    )
-    if row is None:
-        return None
-    out = _engine_row_summary(row)
-    out["scope_kind"] = row["scope_kind"]
-    out["scope_id"] = row["scope_id"]
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def update_procedural_engine(
-    pool: asyncpg.Pool,
-    engine_id: str,
-    *,
-    slug: str,
-    name: str,
-    description: str | None,
-    doc: dict,
-    base_revision: int,
-) -> int | None:
-    """Optimistic-concurrency update of an engine's metadata + manifest doc.
-    Returns the new revision, or None on revision conflict. Propagates
-    asyncpg.UniqueViolationError when the new slug collides in-scope."""
-    row = await pool.fetchrow(
-        """
-        UPDATE procedural_engines
-        SET slug = $2, name = $3, description = $4, doc = $5::jsonb,
-            revision = revision + 1, updated_at = now()
-        WHERE id = $1 AND revision = $6 AND NOT archived
-        RETURNING revision
-        """,
-        engine_id,
-        slug,
-        name,
-        description,
-        json.dumps(doc),
-        base_revision,
-    )
-    return None if row is None else row["revision"]
-
-
-async def get_procedural_engine_by_slug(
-    pool: asyncpg.Pool, *, scope_kind: str, scope_id: str | None, slug: str
-) -> dict | None:
-    """A live engine (with its manifest ``doc``) by slug in a scope — the compile
-    worker resolves a selected external engine's entrypoint/repo this way (the
-    compile request carries only the slug)."""
-    row = await pool.fetchrow(
-        """
-        SELECT id, scope_kind, scope_id, slug, name, description, doc,
-               revision, created_by, created_at, updated_at
-        FROM procedural_engines
-        WHERE scope_kind = $1 AND COALESCE(scope_id, '') = COALESCE($2, '')
-              AND slug = $3 AND NOT archived
-        """,
-        scope_kind,
-        scope_id,
-        slug,
-    )
-    if row is None:
-        return None
-    out = _engine_row_summary(row)
-    out["scope_kind"] = row["scope_kind"]
-    out["scope_id"] = row["scope_id"]
-    out["doc"] = _loads_jsonb(row["doc"])
-    return out
-
-
-async def set_procedural_engine_wheel(pool: asyncpg.Pool, engine_id: str, wheel_key: str | None) -> bool:
-    """Record (or clear) a built engine's wheel pointer in its manifest doc
-    WITHOUT bumping the revision — a background build must not race the user's
-    optimistic commits (mirrors :func:`apply_inferred_bbox`)."""
-    res = await pool.execute(
-        "UPDATE procedural_engines SET doc = jsonb_set(doc, '{wheel_key}', $2::jsonb, true), "
-        "updated_at = now() WHERE id = $1 AND NOT archived",
-        engine_id,
-        json.dumps(wheel_key),
-    )
-    return res.endswith("1")
-
-
-async def archive_procedural_engine(pool: asyncpg.Pool, engine_id: str) -> bool:
-    res = await pool.execute(
-        "UPDATE procedural_engines SET archived = TRUE, updated_at = now() WHERE id = $1 AND NOT archived",
-        engine_id,
-    )
-    return res.endswith("1")
-
-
-# ── source nodes — when each node of an external source last changed ─────────
-#
-# See migrations/028_source_nodes.sql for what the table is for. These helpers
-# are the whole query surface: one write, three reads. Everything a consumer
-# needs is composed from them, and nothing here understands a ref.
-
-
-@dataclass(frozen=True)
-class SourceNode:
-    node_ref: str
-    parent_ref: Optional[str]
-    name: Optional[str]
-    last_changed_at: datetime.datetime
-    last_changed_by: Optional[str]
-    observed_at: datetime.datetime
-
-
-def _source_node_row(r) -> SourceNode:
-    return SourceNode(
-        node_ref=r["node_ref"],
-        parent_ref=r["parent_ref"],
-        name=r["name"],
-        last_changed_at=r["last_changed_at"],
-        last_changed_by=r["last_changed_by"],
-        observed_at=r["observed_at"],
-    )
-
-
-async def record_source_nodes(
-    pool: asyncpg.Pool,
-    *,
-    scope: str,
-    source: str,
-    nodes: list,
-) -> int:
-    """Upsert observed nodes. Returns how many rows were written.
-
-    ``nodes`` is a list of dicts with ``node_ref`` and ``last_changed_at``
-    required, and ``parent_ref`` / ``name`` / ``last_changed_by`` optional.
-
-    LAST_CHANGED_AT ONLY EVER MOVES FORWARD. A writer re-observing a node it has
-    seen before may hold an older cursor than the row does -- a backfill, a
-    re-run over a wider window, two writers with different windows -- and letting
-    that overwrite a newer timestamp would mark current assets stale-free when
-    they are not. `observed_at` is always taken from the new write, because "when
-    was this last confirmed" is precisely the thing that must not be sticky.
-
-    Written as one executemany inside a transaction: an hourly sweep stamps a
-    changed leaf and every node above it, so the natural batch is thousands of
-    rows and a round trip each would dominate.
-    """
-    if not nodes:
-        return 0
-    rows = [
-        (
-            scope,
-            source,
-            n["node_ref"],
-            n.get("parent_ref"),
-            n.get("name"),
-            n["last_changed_at"],
-            n.get("last_changed_by"),
-        )
-        for n in nodes
-    ]
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            await conn.executemany(
-                """
-                INSERT INTO source_nodes
-                    (scope, source, node_ref, parent_ref, name, last_changed_at, last_changed_by)
-                VALUES ($1, $2, $3, $4, $5, $6, $7)
-                ON CONFLICT (scope, source, node_ref) DO UPDATE SET
-                    parent_ref      = COALESCE(EXCLUDED.parent_ref, source_nodes.parent_ref),
-                    name            = COALESCE(EXCLUDED.name, source_nodes.name),
-                    last_changed_at = GREATEST(EXCLUDED.last_changed_at, source_nodes.last_changed_at),
-                    last_changed_by = CASE
-                        WHEN EXCLUDED.last_changed_at >= source_nodes.last_changed_at
-                        THEN EXCLUDED.last_changed_by ELSE source_nodes.last_changed_by END,
-                    observed_at     = NOW()
-                """,
-                rows,
-            )
-    return len(rows)
-
-
-async def get_source_node(pool: asyncpg.Pool, *, scope: str, source: str, node_ref: str) -> Optional[SourceNode]:
-    """One node, or None. The staleness check: compare `last_changed_at` against
-    whatever timestamp the caller's copy was made at."""
-    r = await pool.fetchrow(
-        "SELECT node_ref, parent_ref, name, last_changed_at, last_changed_by, observed_at "
-        "FROM source_nodes WHERE scope = $1 AND source = $2 AND node_ref = $3",
-        scope,
-        source,
-        node_ref,
-    )
-    return _source_node_row(r) if r else None
-
-
-async def get_source_nodes(pool: asyncpg.Pool, *, scope: str, source: str, node_refs: list) -> list:
-    """Many nodes in one round trip.
-
-    The batch form exists because the interesting caller holds a whole listing of
-    published assets and wants every answer at once; asking one at a time turns a
-    freshness check into N round trips over a slow link.
-    """
-    if not node_refs:
-        return []
-    rows = await pool.fetch(
-        "SELECT node_ref, parent_ref, name, last_changed_at, last_changed_by, observed_at "
-        "FROM source_nodes WHERE scope = $1 AND source = $2 AND node_ref = ANY($3::text[])",
-        scope,
-        source,
-        list(node_refs),
-    )
-    return [_source_node_row(r) for r in rows]
-
-
-async def list_source_nodes_changed_since(
-    pool: asyncpg.Pool,
-    *,
-    scope: str,
-    source: str,
-    since: Optional[datetime.datetime] = None,
-    limit: int = 1000,
-) -> list:
-    """Nodes whose last change is newer than ``since``, newest first.
-
-    The polling shape: a consumer that keeps its own cursor asks what has moved
-    rather than re-checking everything it holds. ``limit`` is capped by the
-    caller; the index makes the ordering free.
-    """
-    if since is None:
-        rows = await pool.fetch(
-            "SELECT node_ref, parent_ref, name, last_changed_at, last_changed_by, observed_at "
-            "FROM source_nodes WHERE scope = $1 AND source = $2 "
-            "ORDER BY last_changed_at DESC LIMIT $3",
-            scope,
-            source,
-            limit,
-        )
-    else:
-        rows = await pool.fetch(
-            "SELECT node_ref, parent_ref, name, last_changed_at, last_changed_by, observed_at "
-            "FROM source_nodes WHERE scope = $1 AND source = $2 AND last_changed_at > $3 "
-            "ORDER BY last_changed_at DESC LIMIT $4",
-            scope,
-            source,
-            since,
-            limit,
-        )
-    return [_source_node_row(r) for r in rows]
-
-
-async def list_source_node_sources(pool: asyncpg.Pool, *, scope: str) -> list:
-    """Which sources have recorded anything in this scope, with their row counts
-    and freshest observation. What a consumer reads before it knows any refs."""
-    rows = await pool.fetch(
-        "SELECT source, COUNT(*) AS nodes, MAX(last_changed_at) AS last_changed_at, "
-        "MAX(observed_at) AS observed_at FROM source_nodes WHERE scope = $1 GROUP BY source ORDER BY source",
-        scope,
-    )
-    return [
-        {
-            "source": r["source"],
-            "nodes": r["nodes"],
-            "last_changed_at": r["last_changed_at"],
-            "observed_at": r["observed_at"],
-        }
-        for r in rows
-    ]
-
-
-# ── Plugin-job schedules (cron for plugin jobs) ─────────────────────
-#
-# Deliberately shaped like the audit_schedules helpers above so the two read the
-# same way side by side -- same claim strategy, same skip-reason convention. See
-# migrations/029_plugin_job_schedules.sql for why they are separate tables.
-
-
-def _plugin_job_schedule_row(r) -> dict:
-    """Project a plugin_job_schedules row to its JSON-ready dict shape."""
-    return {
-        "id": str(r["id"]),
-        "name": r["name"],
-        "cron_expr": r["cron_expr"],
-        "scope": r["scope"],
-        "plugin_id": r["plugin_id"],
-        # asyncpg hands JSONB back as a string unless a codec is registered, and
-        # this module registers none. Decoded here so every caller sees a dict
-        # rather than half of them re-parsing it.
-        "options": (json.loads(r["options"]) if isinstance(r["options"], str) else (r["options"] or {})),
-        "capability": r["capability"],
-        "enabled": r["enabled"],
-        "last_fired_at": (r["last_fired_at"].isoformat() if r["last_fired_at"] else None),
-        "next_fire_at": (r["next_fire_at"].isoformat() if r["next_fire_at"] else None),
-        "last_skipped_reason": r["last_skipped_reason"],
-        "last_job_id": r["last_job_id"],
-        "created_at": r["created_at"].isoformat() if r["created_at"] else None,
-        "created_by": r["created_by"],
-        "archived_at": (r["archived_at"].isoformat() if r["archived_at"] else None),
-    }
-
-
-_PLUGIN_JOB_SCHEDULE_COLS = (
-    "id, name, cron_expr, scope, plugin_id, options, capability, enabled, "
-    "last_fired_at, next_fire_at, last_skipped_reason, last_job_id, "
-    "created_at, created_by, archived_at"
-)
-
-
-async def list_plugin_job_schedules(pool: asyncpg.Pool, *, include_archived: bool = False) -> list:
-    where = "" if include_archived else " WHERE archived_at IS NULL"
-    rows = await pool.fetch(
-        f"SELECT {_PLUGIN_JOB_SCHEDULE_COLS} FROM plugin_job_schedules{where} ORDER BY created_at DESC"
-    )
-    return [_plugin_job_schedule_row(r) for r in rows]
-
-
-async def get_plugin_job_schedule(pool: asyncpg.Pool, schedule_id: str) -> Optional[dict]:
-    row = await pool.fetchrow(
-        f"SELECT {_PLUGIN_JOB_SCHEDULE_COLS} FROM plugin_job_schedules WHERE id = $1",
-        schedule_id,
-    )
-    return _plugin_job_schedule_row(row) if row else None
-
-
-async def create_plugin_job_schedule(
-    pool: asyncpg.Pool,
-    *,
-    name: str,
-    cron_expr: str,
-    scope: str,
-    plugin_id: str,
-    options: dict,
-    capability: Optional[str] = None,
-    enabled: bool = True,
-    next_fire_at=None,
-    created_by: Optional[str] = None,
-) -> dict:
-    """Insert one schedule. ``next_fire_at`` is pre-computed by the caller from
-    ``cron_expr``, which is also where the expression is validated."""
-    row = await pool.fetchrow(
-        f"""
-        INSERT INTO plugin_job_schedules
-            (name, cron_expr, scope, plugin_id, options, capability, enabled, next_fire_at, created_by)
-        VALUES ($1, $2, $3, $4, $5::jsonb, $6, $7, $8, $9)
-        RETURNING {_PLUGIN_JOB_SCHEDULE_COLS}
-        """,
-        name,
-        cron_expr,
-        scope,
-        plugin_id,
-        json.dumps(options or {}),
-        capability,
-        enabled,
-        next_fire_at,
-        created_by,
-    )
-    return _plugin_job_schedule_row(row)
-
-
-async def update_plugin_job_schedule(pool: asyncpg.Pool, schedule_id: str, **fields) -> Optional[dict]:
-    """Partial update. Only the named columns move.
-
-    Built as a dynamic SET list rather than one statement per field, but the
-    column names come from a FIXED allow-list -- a caller cannot name a column
-    that is not here, so the interpolation below never carries a caller's string
-    into the SQL.
-    """
-    allowed = (
-        "name",
-        "cron_expr",
-        "scope",
-        "plugin_id",
-        "options",
-        "capability",
-        "enabled",
-        "next_fire_at",
-        "last_fired_at",
-        "last_skipped_reason",
-        "last_job_id",
-    )
-    sets, values = [], []
-    for column in allowed:
-        if column not in fields:
-            continue
-        value = fields[column]
-        if column == "options":
-            sets.append(f"{column} = ${len(values) + 1}::jsonb")
-            values.append(json.dumps(value or {}))
-        else:
-            sets.append(f"{column} = ${len(values) + 1}")
-            values.append(value)
-    if not sets:
-        return await get_plugin_job_schedule(pool, schedule_id)
-    values.append(schedule_id)
-    row = await pool.fetchrow(
-        f"UPDATE plugin_job_schedules SET {', '.join(sets)} "
-        f"WHERE id = ${len(values)} RETURNING {_PLUGIN_JOB_SCHEDULE_COLS}",
-        *values,
-    )
-    return _plugin_job_schedule_row(row) if row else None
-
-
-async def archive_plugin_job_schedule(pool: asyncpg.Pool, schedule_id: str) -> bool:
-    """Soft delete. The audit rows a schedule produced outlive it, and archiving
-    frees the name for re-use."""
-    result = await pool.execute(
-        "UPDATE plugin_job_schedules SET archived_at = NOW(), enabled = FALSE " "WHERE id = $1 AND archived_at IS NULL",
-        schedule_id,
-    )
-    return result.endswith(" 1")
-
-
-async def claim_due_plugin_job_schedule(pool: asyncpg.Pool, *, now, next_fire_at) -> Optional[dict]:
-    """Atomically claim ONE due schedule, or None.
-
-    ``FOR UPDATE SKIP LOCKED`` so several API replicas can tick concurrently and
-    still fire a given row at most once -- the losers' UPDATE matches nothing and
-    they move on. ``last_skipped_reason`` is cleared on claim; the caller re-sets
-    it if the dispatch then decides not to fire.
-
-    ``next_fire_at`` is a provisional value the caller corrects immediately after,
-    once the row's own ``cron_expr`` is known. Claiming first and computing second
-    is what keeps the claim a single statement.
-    """
-    row = await pool.fetchrow(
-        f"""
-        UPDATE plugin_job_schedules
-        SET last_fired_at = $1,
-            next_fire_at = $2,
-            last_skipped_reason = NULL
-        WHERE id = (
-            SELECT id FROM plugin_job_schedules
-            WHERE enabled
-              AND archived_at IS NULL
-              AND next_fire_at IS NOT NULL
-              AND next_fire_at <= $1
-            ORDER BY next_fire_at ASC
-            LIMIT 1
-            FOR UPDATE SKIP LOCKED
-        )
-        RETURNING {_PLUGIN_JOB_SCHEDULE_COLS}
-        """,
-        now,
-        next_fire_at,
-    )
-    return _plugin_job_schedule_row(row) if row else None
-
-
-async def set_plugin_job_schedule_skip_reason(pool: asyncpg.Pool, schedule_id: str, reason: str) -> None:
-    """Record why a due slot produced no job.
-
-    Best-effort: a schedule that fires matters more than the note explaining one
-    that did not, so this never raises into the tick.
-    """
-    try:
-        await pool.execute(
-            "UPDATE plugin_job_schedules SET last_skipped_reason = $2 WHERE id = $1",
-            schedule_id,
-            reason,
-        )
-    except Exception:
-        logger.exception("plugin-job scheduler: could not record skip reason for %s", schedule_id)
-
-
-async def plugin_job_in_flight_jobs(pool: asyncpg.Pool, *, scope_kind: str, scope_id, plugin_id: str) -> list[str]:
-    """Job ids for this plugin and scope whose audit row is still queued or running.
-
-    The concurrent-fire guard's first pass, and the guard matters more for a plugin
-    than for an audit sweep: a plugin job can hold a single licensed workstation
-    for minutes, so two overlapping firings do not merely double the load -- they
-    contend for one resource, and the loser tends to fail in a way that reads as
-    the plugin's fault rather than as a scheduling one.
-
-    IDS, NOT A BOOLEAN, because this row alone cannot answer the question. The
-    terminal status is written by the WORKER, so a worker with no database pool
-    cannot write it -- it says as much at startup -- and in that deployment every
-    plugin job stays `queued` here for good. A boolean would then report a finished
-    job as still in flight and block its schedule permanently after the first
-    firing, which is a worse failure than the double-firing it set out to prevent.
-    The caller checks these ids against the queue, which the worker does update.
-
-    Matched on the synthetic source key's prefix, which is where the plugin id
-    lives; audit_log carries no plugin column. ``starts_with`` rather than
-    ``LIKE``: that key begins with an underscore, which LIKE reads as a
-    single-character wildcard.
-
-    Rows with no ``job_id`` are skipped: there is nothing to check them against,
-    and one is not evidence of a running job -- it is an audit row that never got
-    as far as being queued.
-    """
-    prefix = f"_synthetic/plugin_job/{plugin_id}/"
-    rows = await pool.fetch(
-        """
-        SELECT job_id FROM audit_log
-        WHERE target_format = 'plugin_job'
-          AND status IN ('queued', 'running')
-          AND scope_kind = $1
-          AND scope_id IS NOT DISTINCT FROM $2
-          AND starts_with(key, $3)
-          AND job_id IS NOT NULL
-        """,
-        scope_kind,
-        scope_id,
-        prefix,
-    )
-    return [str(r["job_id"]) for r in rows]
